% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesize.R
\name{add_data}
\alias{add_data}
\alias{add_data.hypothesis}
\alias{add_data.framework}
\title{Adding Data}
\usage{
add_data(x, ...)

\method{add_data}{hypothesis}(x, data, data_name = deparse(substitute(data)))

\method{add_data}{framework}(x, hypothesis_name, data, data_name = deparse(substitute(data)))
}
\arguments{
\item{x}{Either a \code{hypothesis} or \code{framework} object}

\item{data}{Data set to be attached to this hypothesis. Not necessary to
create a \code{hypothesis}, and can be specified or added at a later time. The
name of the data set will be used as the reference variable.}

\item{data_name}{Name of the dataset that this hypothesis should be tested
against. The data itself will be used to generate the name if it is given
from the \emph{data} argument. If only a character vector is given, the function
will match the \code{hypothesis} to data already available (or warn if not
available). This flexibility allows for lightweight manipulation of the
\code{hypothesis} object, and the addition of multiple data sets.}

\item{hypothesis_name}{Character vector of which hypothesis data frame should be added to. This will replace prior data that was set.}

\item{strata}{How the data should be split or stratified. References the
name of the data given in that the models will be with fit
against, splitting the data into subsets. This helps to perform hypothesis
testing on subsets or strata of the data. It defaults to NULL (which means
the full data will be used) \strong{experimental}}
}
\description{
If a hypothesis does not have a data, it can be added subsequently. This can be done in a pipe, either to the hypothesis itself, or to a framework, to allow for flexibility.
}
\concept{helpers}
