% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aim.R
\name{aim}
\alias{aim}
\title{Aiming For Target Hypotheses}
\usage{
aim(bullet, ...)
}
\arguments{
\item{bullet}{Bullet object that is a hypothesis with permutations of how a
model is built, based on the statistical plan.}

\item{...}{For extensibility}
}
\value{
An \code{aim} object that contain all the relevant hypothesis that were
pre-specified, including model specifications from \code{tidymodels}, in a table
format.
}
\description{
This creates the hypothesis that are being prepared for
analysis. These contain model specifications and the pipeline, along with
model identifiers.
}
\examples{
library(parsnip)
f <- mpg + cyl ~ wt + hp + gear
lm_mod <- linear_reg() \%>\% set_engine("lm")
b <- bullet(f, exposure = "wt", approach = "sequential", model = lm_mod)
a <- aim(b)

}
