% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aims-package.R
\docType{package}
\name{aims-package}
\alias{aims}
\alias{aims-package}
\title{aims: A Many-Handed Approach to Multiple Models}
\description{
A simple and tidy way to organize the many models built during a 
    research project. The package consists of individual hypothesis objects that 
    are all linked together around a single data set. In causality-based 
    research, such as epidemiology, the focus is on specific hypothesis that can
    be tested. However, the number of models can become unruly. This serves as a 
    way to organize the many.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/asshah4/aims}
  \item Report bugs at \url{https://github.com/asshah4/aims/issues}
}

}
\author{
\strong{Maintainer}: Anish Shah \email{ashah282@uic.edu} (\href{https://orcid.org/0000-0002-9729-1558}{ORCID})

}
\keyword{internal}
