% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-table.R
\name{attach_data}
\alias{attach_data}
\title{Attach a dataset to a model table}
\usage{
attach_data(x, data, ...)
}
\arguments{
\item{x}{a \verb{<mdl_tbl>} object}

\item{data}{a \verb{<data.frame>} object that has been used by models}
}
\description{
Attach a dataset to a model table
}
\details{
When models are built, oftentimes the included matrix of data is available
within the raw model, however when handling many models, this can be
expensive in terms of memory and space. By attaching datasets independently
that persist regardless of the underlying models, and by knowing which models
used which datasets, it can be ease to back-transform information.
}
