% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{augment.fmls}
\alias{augment.fmls}
\title{Augment a formula with information from a \code{fmls} object}
\usage{
\method{augment}{fmls}(x, ...)
}
\description{
Augment a formula with information from a \code{fmls} object
}
