% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ballistics.R
\name{ballistics}
\alias{ballistics}
\alias{ballistics.default}
\alias{ballistics.bullet}
\alias{ballistics.aim}
\title{Ballistic Analysis of the Hypotheses}
\usage{
ballistics(mark, ...)

\method{ballistics}{default}(mark, ...)

\method{ballistics}{bullet}(mark, ...)

\method{ballistics}{aim}(mark, ...)
}
\arguments{
\item{mark}{Represents either a \code{bullet} or \code{aim} object to be analyzed for
further information or characteristics
\itemize{
\item A single \strong{bullet} object will return a table that is structured for the
proposed analytical approach
\item A single \strong{aim} object, after the \code{fire()} function, will return a table
that includes further analytical elements. Currently returns the runtime in
miliseconds.
}}

\item{...}{For extensibility}
}
\value{
The function returns values based on what object was given to it.

A \code{bullet} object returns a tibble with the following columns:
\itemize{
\item outcomes: character vector
\item model_num: an identifying model number
\item formulas: a list of formulas that will be used for modeling
\item model_spec: the \code{parsnip} model specification
}

An \code{aim} object can be returned in two different ways.
\itemize{
\item If it has not yet been processed, it returns a table with additional
columns identifying potential problems before it can be processed by the
\code{fire()} function
\item If it has been processed by the \code{fire()} function, it will return
additional columns to the original table that include model statistics,
including a tidied fitted object
}
}
\description{
Look at the overall characteristics of the hypotheses, such as
number of regression models and how they will be put together based on the
approach. Is used to evaluate the objects created by \code{bullet()} and by
\code{aim()} (which also uses ballistics to generate the appropriate modeling
matrix).
}
\examples{
library(parsnip)
f <- mpg + cyl ~ wt + hp + gear
lm_mod <- linear_reg() \%>\% set_engine("lm")
b <- bullet(f, exposure = "wt", approach = "parallel", model = lm_mod)
ballistics(b)
}
\concept{ballistics}
