% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framework.R
\name{build_frames}
\alias{build_frames}
\title{Build Up the Framework by Fitting Models and Tests}
\usage{
build_frames(framework, which_ones = NULL, ...)
}
\arguments{
\item{framework}{Object of class \code{framework}}

\item{which_ones}{Vector of which hypothesis should be built. Defaults to
building all hypotheses that have not been run yet.}
}
\value{
Returns a \code{framework} object that has be fitted and tidied.
}
\description{
This function allows for delayed building of multiple hypothesis. It uses the
\code{hypothesis} objects with the corresponding \strong{test} arguments against the
prescribed \strong{data}. Which hypotheses to run can be specified, which will
forcibly re-run these, otherwise the default behavior is to only run models
that have not yet been fitted.
}
\seealso{
Other frameworks: 
\code{\link{add_hypothesis}()},
\code{\link{fit_models}()},
\code{\link{framework}()}
}
\concept{frameworks}
