% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_models.R
\name{build_models}
\alias{build_models}
\title{Build the Models and Tests}
\usage{
build_models(project, stage = "build", ...)
}
\arguments{
\item{project}{Object of class \code{project}}

\item{stage}{Internal marker of workflow progress.}

\item{...}{For extensibility}
}
\value{
Object of \code{project} class
}
\description{
The project arms may include hypotheses that are represented by multiple
types of tests. The models are dependent on the universal definitions from
the \href{https://parsnip.tidymodels.org/}{{parsnip}} package. In addition,
specific basic statistical testing can be analyzed and stored.
This step is added after \code{add_hypothesis()} in terms of workflow. This
function only runs on hypothesis that have not already been run, allowing for
the ease of adding to an existing project without rebuilding it.
}
