% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{build_study}
\alias{build_study}
\title{Run the Study by Fitting Models and Tests}
\usage{
build_study(study, which_ones = NULL, ...)
}
\arguments{
\item{study}{Object of class \code{study}}

\item{which_ones}{Vector of which hypothesis should be built. Defaults to
building all hypotheses that have not been run yet. If given a name, will
forcibly re-run the analysis.}
}
\value{
Returns a \code{study} object that has be fitted and tidied.
}
\description{
This function allows for delayed building of multiple hypothesis. It uses the
\code{hypothesis} objects with the corresponding \strong{test} arguments against the
prescribed \strong{data}. Which hypotheses to run can be specified, which will
forcibly re-run these, otherwise the default behavior is to only run models
that have not yet been fitted.
}
\seealso{
Other studies: 
\code{\link{add_hypothesis}()},
\code{\link{study}()}
}
\concept{studies}
