% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-table.R
\name{can_reconstruct_model_table}
\alias{can_reconstruct_model_table}
\title{Can \code{md_tbl} be reconstructed based on invariants?}
\usage{
can_reconstruct_model_table(x, to)
}
\arguments{
\item{x}{data frame that will have invariants checked}

\item{to}{the tibble subclass of \code{md_tbl} that would be reconstructed}
}
\description{
Can \code{md_tbl} be reconstructed based on invariants?
}
\keyword{internal}
