% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carve.R
\name{carve_out}
\alias{carve_out}
\title{Carve Out Confounders}
\usage{
carve_out(study, name, ...)
}
\arguments{
\item{study}{A \code{study} object that has been constructed}

\item{name}{Name of a \code{hypothesis} object}

\item{...}{For extensibility}
}
\description{
Carve Out Confounders
}
