% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity.R
\name{complexity}
\alias{complexity}
\title{Identify complexity of a formula}
\usage{
complexity(x)
}
\arguments{
\item{x}{An object that is either a single \code{formula}, \code{fmls} or vector of \code{tm} objects}
}
\value{
A single integery value from \code{0:5}
}
\description{
A \code{formula} or \code{fmls} object may have several levels of complexity that need
to be assessed. This affects how the formula may need to be expanded based on
specific or special terms, such as a mediation, interaction, etc. It as
designed for usage with the built-in \code{fmls} class and its corresponding \code{tm}
objects, however, can be used for generic formulas if needed as well.
}
\section{Rules of Roles}{
Roles may be given or assumed based on the formula definition, such as LHS
objects being outcomes and RHS objects being predictors, particularly in a
model-focused paradigm. This makes certain assumptions about the prescribed
roles that a term can be given:
\enumerate{
\item There can only be a single \code{outcome} term on the \emph{LHS}.
\item The \emph{RHS} can not contain \code{outcome} terms.
\item There can only be a single \code{exposure} term on the \emph{RHS}.
\item \code{interaction} terms must be expanded against an \code{exposure}, such that both
terms are present independently and as an interaction e.g. \code{Y ~ X + I + X:I}
\item There may not be > 1 \code{mediator} on the \emph{RHS}
\item \code{mediator} terms on the \emph{RHS} may either exist alone, or alongside an \code{exposure} term, but not just with other \code{confounder} terms (as they can be assessed against the \code{outcome} directly as an adjustment for the \code{exposure}).
\item \code{mediator} terms may exist on the \emph{LHS}, but the \code{exposure} term must be oon the \emph{RHS} without other \code{confounder} terms.
\item A single \emph{meta} term, such as \code{strata}, may only exist on the \emph{RHS}.
}
}

\section{Complexity}{
The complexity of a formula is derived from how well it follows the \emph{rules of
roles}, how informative the \emph{directionality} of the terms are, and the number
of terms on the \emph{LHS} and \emph{RHS}\tabular{lccrr}{
   Order \tab Rules \tab Directionality \tab LHS \tab RHS \cr
   0th \tab + \tab - \tab ≤ 1 \tab ≤ 1 \cr
   1st \tab - \tab - \tab 1 \tab 1 \cr
   2nd \tab + \tab + \tab 1 \tab ≥ 1 \cr
   3rd \tab - \tab - \tab 1 \tab ≥ 1 \cr
   4th \tab - \tab + \tab > 1 \tab ≥ 1 \cr
}
}

\seealso{
\code{\link[=tm]{tm()}} for review of roles that can be defined.
}
