% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms.R
\name{components.terms}
\alias{components.terms}
\title{Extract components from a vector of terms (\code{tm})}
\usage{
\method{components}{terms}(x, ...)
}
\arguments{
\item{x}{A vector of \code{tm} objects}

\item{...}{A series of \code{name = value} pairs that represent the attribute to
retrieve. Can have a value of \strong{\if{html}{\out{<NA>}}} if needed to find missing attributes
of terms. This is a filtering process, so all requirements will be returned}
}
\value{
A \code{tm} object
}
\description{
This is a \code{\link[generics:components]{generics::components()}} method for extract individual components from a formula.
}
