% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{construct_dag}
\alias{construct_dag}
\title{Construct Directed Acyclic Graphs}
\usage{
construct_dag(study, name, tidy = FALSE)
}
\arguments{
\item{study}{Object of class \code{study}}

\item{name}{Name of the \code{hypothesis} object, which defaults to the name of
the \code{hypothesis} object itself}

\item{tidy}{Defaults to FALSE, thus returning a \code{dagitty} object. If TRUE, then will return a \code{tidy_dagitty} object.}
}
\value{
\code{dagitty} or \code{tidy_dagitty} object
}
\description{
This function converts a hypothesis into a \code{dagitty} object (or
\code{tidy_dagitty} if requested). This can subsequently be passed onto the
\code{\link[ggdag:ggdag]{ggdag::ggdag()}} function for additional plotting.
}
