% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-table.R
\name{construct_formula_table}
\alias{construct_formula_table}
\title{Restructure formulas to fit within a model table}
\usage{
construct_formula_table(x, ...)
}
\arguments{
\item{x}{Vector of \code{fmls} objects}
}
\description{
Restructure formulas to fit within a model table
}
\keyword{internal}
