% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{construct_map}
\alias{construct_map}
\title{Construct Map}
\usage{
construct_map(study, which_ones = NULL, ...)
}
\arguments{
\item{study}{A \code{study} object that contains \code{hypothesis} objects}

\item{which_ones}{Vector of which hypothesis should be constructed. It
defaults to NULL, which constructs all hypotheses that have not yet been
processed. If a vector of names is given, will forcibly re-analyze them.}

\item{...}{For extensibility}
}
\value{
Invisibly returns a \code{study} object that has the hypotheses mapped to
it, including fits and paths
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function allows for delayed building of multiple hypothesis. It uses the
\code{hypothesis} objects with the corresponding \strong{test} arguments against the
prescribed \strong{data}. Which hypotheses to run can be specified, which will
forcibly re-run these, otherwise the default behavior is to only run models
that have not yet been fitted.
}
\seealso{
Other studies: 
\code{\link{create_study}()}
}
\concept{studies}
