% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-table.R
\name{construct_model_table}
\alias{construct_model_table}
\title{Restructure models to fit within a model table}
\usage{
construct_model_table(x, ...)
}
\arguments{
\item{x}{Vector of \code{mdl} objects}
}
\description{
Restructure models to fit within a model table
}
\keyword{internal}
