% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{construct_models}
\alias{construct_models}
\title{Construct Models}
\usage{
construct_models(study, which_ones = NULL, ...)
}
\arguments{
\item{study}{A \code{study} object that contains \code{hypothesis} objects}

\item{which_ones}{Vector of which hypothesis should be constructed. It
defaults to NULL, which constructs all hypotheses that have not yet been
processed. If a vector of names is given, will forcibly re-analyze them.}

\item{...}{For extensibility}
}
\value{
Invisibly returns a \code{study} object that has be fitted and tidied
}
\description{
This function allows for delayed building of multiple hypothesis. It uses the
\code{hypothesis} objects with the corresponding \strong{test} arguments against the
prescribed \strong{data}. Which hypotheses to run can be specified, which will
forcibly re-run these, otherwise the default behavior is to only run models
that have not yet been fitted.
}
\seealso{
Other studies: 
\code{\link{construct_paths}()},
\code{\link{draw_hypothesis}()},
\code{\link{study}()}
}
\concept{studies}
