% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{construct_paths}
\alias{construct_paths}
\title{Construct Paths}
\usage{
construct_paths(study, which_ones = NULL, ...)
}
\arguments{
\item{study}{A \code{study} object that contains \code{hypothesis} objects}

\item{which_ones}{Vector of which hypothesis should be constructed. It
defaults to NULL, which constructs all hypotheses that have not yet been
processed. If a vector of names is given, will forcibly re-analyze them.}

\item{...}{For extensibility}
}
\value{
A \code{study} object with paths added
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Takes the \code{hypothesis} objects that were added to the \code{study} and decomposes
them into specific \strong{paths} that are used to help define relationships
between variables. These paths are stored in the \code{study} itself in the form
of a \code{data.frame}. This is currently experimental in that the directionality,
relationships, and patterns are intended to be used to help identify certain
variables for future modeling, but the implementation is not yet complete.
}
\seealso{
Other studies: 
\code{\link{construct_models}()},
\code{\link{draw_hypothesis}()},
\code{\link{study}()}
}
\concept{studies}
