% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{create_map}
\alias{create_map}
\title{Mapping Many Hypotheses Together}
\usage{
create_map(...)
}
\arguments{
\item{...}{For extensibility}
}
\value{
A \code{model_map} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Calling \code{create_map()} initializes a list object that stores \code{hypothesis} objects,
which are used to explore a research project. It allows for delayed building
of models, and is used to help study the relationship between variables.

A \code{model_map} object is a essentially a modified list containing two
complementary table structures, which are interrelated via the \code{hypothesis}
objects, and several attributes that allow for information storage about the
object itself.
}
\concept{maps}
