% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagger-package.R
\docType{package}
\name{dagger-package}
\alias{dagger}
\alias{dagger-package}
\title{dagger: Directed Acyclic Graphs Generated by Exploratory Research}
\description{
A simple and tidy way to organize the many models built during a 
    research project. The package consists of formulating hypothesis, testing or
    modeling the findings, and understanding the parameters. In causality-based 
    research, such as epidemiology, the focus is on a specific aim that 
    can be tested, which helps to contextualize variables, and provides 
    relationships between outcomes, exposures, and covariates. This package 
    attempts to provide a framework for these types of analyses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/asshah4/dagger}
  \item Report bugs at \url{https://github.com/asshah4/dagger/issues}
}

}
\author{
\strong{Maintainer}: Anish Shah \email{ashah282@uic.edu} (\href{https://orcid.org/0000-0002-9729-1558}{ORCID})

}
\keyword{internal}
