% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms.R
\name{describe}
\alias{describe}
\title{Describe attributes of a \code{tm} vector}
\usage{
describe(x, property)
}
\arguments{
\item{x}{A vector \code{tm} objects}

\item{property}{A character vector of the following attributes of a \code{tm}
object: role, side, label, group, description, type, distribution}
}
\description{
Describe attributes of a \code{tm} vector
}
