% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.link_hypothesis_to_data}
\alias{.link_hypothesis_to_data}
\title{Link Hypothesis to Data}
\usage{
.link_hypothesis_to_data(framework, hypothesis, name)
}
\description{
Sets the status of the \code{framework} object to tie the hypothesis to the data
it will be tested against.
}
\keyword{internal}
