% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.set_hypothesis_data}
\alias{.set_hypothesis_data}
\title{Set Data in a Hypothesis}
\usage{
.set_hypothesis_data(hypothesis, .data, .data_name, .strata)
}
\description{
Add data to a formula (internal to the \code{hypothesize()} function), along with
potential stratification instructions. Modifies the formula by adding the
attributes to the formula. Data argument can either be data frame or a
character vector referring to a data set that will be added later.
}
\keyword{internal}
