% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.set_hypothesis_formulae}
\alias{.set_hypothesis_formulae}
\title{Set Formulas in a Hypothesis}
\usage{
.set_hypothesis_formulae(hypothesis, formula, combination)
}
\description{
Expand the formula based on outcomes and predictors (including both
exposures, covariates, and fixed variables). The pattern is decided by the
\code{combination} argument. The major functionality is within the
\code{expand_formula()} function, but this allows for setting internal attributes.
}
\keyword{internal}
