% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.set_hypothesis_tests}
\alias{.set_hypothesis_tests}
\title{Set Test and Test Options in a Hypothesis}
\usage{
.set_hypothesis_tests(hypothesis, test, .test_opts)
}
\description{
Add the test and test options (if needed) the the \code{hypothesis} object. This modifies the attributes of the original object and invisibly returns the same ojbect.
}
\keyword{internal}
