% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw}
\alias{draw}
\title{Draw out the map}
\usage{
draw(...)
}
\arguments{
\item{...}{For extensibility}
}
\value{
a \code{ggraph} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This experimental function is to integrate with \code{\link[ggraph:ggraph]{ggraph::ggraph()}} to help look at path relationships in the variables that are described. It may start to resemble a \emph{DAG} as the program develops.
}
