% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorMetric.R
\name{errorMetric}
\alias{errorMetric}
\title{Calculate Error Metrics for Predictions}
\usage{
errorMetric(actual, predicted, metric = "MSE")
}
\arguments{
\item{actual}{A numeric vector of actual (observed) values.}

\item{predicted}{A numeric vector of predicted values.}

\item{metric}{Character indicating the type of error metric to return.
If \code{metric = 'MSE'} (default), the function returns the Mean Squared Error (MSE).
If \code{metric = 'RMSE'}, it returns the Root Mean Squared Error (RMSE).}
}
\value{
A numeric value representing the error metric.
If \code{metric = 'MSE'},  it returns the Mean Squared Error;
If \code{metric = 'RMSE'}, it returns the Root Mean Squared Error.
}
\description{
This function computes error metrics between actual and predicted values.
It can return either the Mean Squared Error (MSE) or the Root Mean Squared Error (RMSE).
}
\examples{
actual <- c(1, 2, 3, 4, 5)
predicted <- c(1.1, 1.9, 3.1, 3.9, 4.8)
errorMetric(actual, predicted) # Returns MSE
errorMetric(actual, predicted, metric = "RMSE") # Returns RMSE

}
