% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis-helpers.R
\name{expand_formula}
\alias{expand_formula}
\title{Expand Formulae for a Hypothesis}
\usage{
expand_formula(formula, labels = NULL, combination, table = FALSE, ...)
}
\arguments{
\item{formula}{Formula showing relationship of outcomes and predictors, and
is essentially the hypothesis. It allows for complex formulae e.g. multiple
predictors, multiple exposures, and multiple covariates, that can be
organized into individual formulae based on additional modifiers. Please
see the  for further documentation.}

\item{labels}{A named list pair of term labels and a character vector of
terms, such as as \code{list(exposure = "x1", outcome = "y")}}

\item{combination}{The building pattern for how to put together the overall
plan. It defines variable relationships that will be used. The options of
\code{combination} will lead to the following patterns:
\itemize{
\item \code{direct} will not modify the original equation pattern
}

\deqn{y ~ x1 + x2}
\itemize{
\item \code{sequential} will create a list of sequentially adjusted formula
}

\deqn{y ~ x1} \deqn{y ~ x1 + x2}
\itemize{
\item \code{parallel} will create a list of formula with parallel predictors
}

\deqn{y ~ x1} \deqn{y ~ x2}}

\item{table}{Determines if the returned value should be a list of formulas or
a table. Defaults to FALSE, leading to a list generation.}

\item{...}{Not currently used}
}
\value{
List or table of formulae and extracted terms (to be determined by
the \strong{table} argument).
}
\description{
This is an internal function that helps to create the appropriate formula
combinations for adding hypotheses.
}
\section{Formulae}{


For each \code{formula}, the terms in the RHS and LHS can be labeled. Each individual LHS term is considered an individual outcome, and each RHS is a predictor which can be grouped in a variety of ways. If labels are provided, then the \code{formula} will be expanded in a certain pattern. The key labels are:
\itemize{
\item \strong{exposures} = defines an independent exposure, which will be placed in a separate formulae from any other term marked as such
\item \strong{fixed} = defines any predictor that should be maintained/fixed in all models, including complex terms such as mixed/random effects
}
}

