% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{expand_paths}
\alias{expand_paths}
\title{Expand Paths of a Formula}
\usage{
expand_paths(formula, ...)
}
\arguments{
\item{formula}{an object of class \code{formula}}

\item{...}{For extensibility}
}
\description{
This presumes that all covariates are confounders are potential confounders.
}
