% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{expand_patterns}
\alias{expand_patterns}
\title{Expand formula by a pattern
Must occur after the formula has been simplified}
\usage{
expand_patterns(x, ...)
}
\description{
Expand formula by a pattern
Must occur after the formula has been simplified
}
