% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\title{Extract a Model Hypothesis from a Study}
\usage{
extract(study, which_ones = NULL, tidy = TRUE, ...)
}
\arguments{
\item{study}{A \code{study} object that contains \code{hypothesis} objects that have
been run}

\item{which_ones}{Vector of names of hypotheses that should be extracted.
Will only return models that have been fit. Defaults to NULL, for which all
tests will be returned.}

\item{tidy}{Logical value for if the returned models should be collected into
a tidy table or as a list of raw model/test fits. Defaults to TRUE.}

\item{...}{For extensibility}
}
\value{
A table of model/test fits, with either raw or tidied results
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Extracts the model components of a hypothesis that has been ran. Everything
is returned as a table, with a column of either tidy or raw results.
}
