% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{extract_dagitty}
\alias{extract_dagitty}
\title{Extract a \code{dagitty} object}
\usage{
extract_dagitty(study, name, tidy = FALSE)
}
\arguments{
\item{study}{A \code{study} object}

\item{name}{The name of a hypothesis added to the study}

\item{tidy}{Defaults to FALSE, thus returning a \code{dagitty} object. If TRUE,
then will return a \code{tidy_dagitty} object.}
}
\value{
\code{dagitty} or \code{tidy_dagitty} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function converts a hypothesis into a \code{dagitty} object (or
\code{tidy_dagitty} if requested). This can subsequently be passed onto the
\code{\link[ggdag:ggdag]{ggdag::ggdag()}} function for additional plotting.
}
\concept{extractors visualizers}
