% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_results}
\alias{extract_results}
\title{Extract models from a study}
\usage{
extract_results(model_map, which_ones = NULL, how = "fit", flat = FALSE, ...)
}
\arguments{
\item{model_map}{A \code{model_map} object that contains \code{hypothesis} objects that have
been run}

\item{which_ones}{Vector of names of hypotheses that should be extracted and
distilled. Will only be applied to models that have been fit. Defaults to
NULL, for which all tests will be returned.}

\item{how}{A character string that describes how the test results should be
returned. The choices are as follows:
\itemize{
\item \strong{fit} = unmodified, raw model or test fit, which is the default option
\item \strong{tidy} = tidy table of parameter estimates, confidence interval, and
test statistics, using the \code{\link[broom:reexports]{broom::tidy()}} approach for the exported method
\item \strong{glance} = a single-row summary of overall model or test as defined by
the exported method from \code{\link[broom:reexports]{broom::tidy()}}
}}

\item{flat}{Logical value for if the findings should be returned as a
simple, unnested table or if the extracted values should be returned in a
column as part of the original \code{model_map} object. Defaults to FALSE. This
allows the function to more commonly be used within a pipe.}

\item{...}{For additional parameters based on above method listed under the
\strong{how} parameter, such as \code{conf.level = 0.95} to be passed along when
performing a tidying function}
}
\value{
Returns either a \code{model_map} object or a \code{tibble} based on the
\strong{flat} argument
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function serves as a non-pipeable extraction function to obtain
aggregate components from a \code{model_map} object. Its designed to have a simple
call interface to allow for retrieval of model objects to be formatted and
displayed.
}
\seealso{
Other extractors: 
\code{\link{flatten}()}
}
\concept{extractors}
