% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-helpers.R
\name{fetch}
\alias{fetch}
\alias{fetch_data}
\alias{fetch_data_name}
\alias{fetch_strata}
\alias{fetch_test}
\alias{fetch_test_opts}
\alias{fetch_combination}
\alias{fetch_formulae}
\alias{fetch_call}
\alias{fetch_tidy}
\alias{fetch_raw}
\alias{fetch_names}
\title{Retrieve, Modify, and Destroy Components From \code{study} Object}
\usage{
fetch_data(x, name)

fetch_data_name(x, name)

fetch_strata(x, name)

fetch_test(x, name)

fetch_test_opts(x, name)

fetch_combination(x, name)

fetch_formulae(x, name)

fetch_call(x, name)

fetch_tidy(x, name)

fetch_raw(x, name)

fetch_names(x, ...)
}
\arguments{
\item{x}{\code{study} object}

\item{name}{Name of \code{hypothesis} object to pull components from}

\item{...}{For additional parameters to be passed on}
}
\value{
The \verb{fetch_*()} functions return the named object from the \code{study}.
The \verb{update_*()} functions return the entire study after modification.
}
\description{
These functions are essentially helper functions to simplify the manipulation
of \code{study} objects. They may retrieve or modify or destroy components of
the \code{study} object as described.

\strong{Retrieval functions}:
\itemize{
\item \code{fetch_data()} retrieves the data set associated with a specified hypothesis
and data name
\item \code{fetch_test()} retrieves the test associated with a hypothesis
\item \code{fetch_formulae()} retrieves the list of formulas associated with a
hypothesis
\item \code{fetch_combination()} retrieves how a hypothesis was combined to generate its
formula list
\item \code{fetch_names()} retrieves the hypothesis names that match a certain argument, such as \code{combination = "parallel"} or \code{data_name = "mtcars"}
}

\strong{Update functions}:
\itemize{
\item \code{update_study()} modifies the components of a study, usually used
internally to help update the stage/status of the specified object
}
}
\concept{fetchers}
