% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confound.R
\name{find_significant}
\alias{find_significant}
\title{Derive Formula with Significant Predictors}
\usage{
find_significant(
  study,
  name,
  statistic = "p.value",
  threshold = 0.05,
  return_study = FALSE,
  ...
)
}
\description{
Derive Formula with Significant Predictors
}
