% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms.R
\name{find_terms}
\alias{find_terms}
\title{Find matching terms by criteria}
\usage{
find_terms(x, ...)
}
\arguments{
\item{x}{A vector of \code{tm} objects}

\item{...}{A series of \code{name = value} pairs that represent the attribute to
retrieve. Can have a value of \strong{\if{html}{\out{<NA>}}} if needed to find missing attributes
of terms. This is a filtering process, so all requirements will be returned}
}
\value{
A \code{tm} object
}
\description{
Find matching terms by criteria
}
