% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit_calls}
\alias{fit_calls}
\title{Fit a list of \code{htest} objects}
\usage{
fit_calls(.formula, .test, .opts = NULL, .data)
}
\arguments{
\item{.formula}{List or vector of formulas}

\item{.test}{An \code{htest} to be called}

\item{.opts}{Options to pass to the test, if needed}

\item{.data}{Data set to be used}
}
\value{
Returns a list of \code{htest} objects
}
\description{
This allows for the simple testing of multiple h-test objects.
It requires three components: a formula, a {parsnip} model definition, and
a data set to analyze.
}
\seealso{
Other helpers: 
\code{\link{fit_parsnip_models}()},
\code{\link{tidy_models}()}
}
\concept{helpers}
