% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_models.R, R/utils.R
\name{fit_models}
\alias{fit_models}
\title{Fit a list of {parsnip} models}
\usage{
fit_models(.formula, .test, .opts = NULL, .data)

fit_models(.formula, .test, .opts = NULL, .data)
}
\arguments{
\item{.formula}{List or vector of formulas}

\item{.test}{A model definition from {parsnip}}

\item{.opts}{Options to pass to the test, if needed}

\item{.data}{Data set to be used}
}
\value{
Returns a list of model fits

Returns a list of model fits
}
\description{
This allows for the simple fitting of multiple models. It
requires three components: a formula, a {parsnip} model definition, and a
data set to analyze.

This allows for the simple fitting of multiple models. It
requires three components: a formula, a {parsnip} model definition, and a
data set to analyze.
}
\seealso{
Other frameworks: 
\code{\link{add_hypothesis}()},
\code{\link{build_frames}()},
\code{\link{framework}()}

Other frameworks: 
\code{\link{add_hypothesis}()},
\code{\link{build_frames}()},
\code{\link{framework}()}
}
\concept{frameworks}
