% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit_parsnip_models}
\alias{fit_parsnip_models}
\title{Fit a list of {parsnip} models}
\usage{
fit_parsnip_models(.formula, .test, .opts = NULL, .data)
}
\arguments{
\item{.formula}{List or vector of formulas}

\item{.test}{A model definition from {parsnip}}

\item{.opts}{Options to pass to the test, if needed}

\item{.data}{Data set to be used}
}
\value{
Returns a list of model fits
}
\description{
This allows for the simple fitting of multiple models. It
requires three components: a formula, a {parsnip} model definition, and a
data set to analyze.
}
\seealso{
Other helpers: 
\code{\link{fit_calls}()}
}
\concept{helpers}
