% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flatten}
\alias{flatten}
\title{Flatten a Model Map}
\usage{
flatten(model_map, check = TRUE)
}
\arguments{
\item{model_map}{A \code{model_map} object that contains \code{hypothesis} objects that have
been run}

\item{check}{Internal usage to turn off validation for work inside other functions}
}
\value{
Returns a \code{tibble} object, without the attributes/features of a
\code{model_map} object
}
\description{
Takes a model that has had results extracted in nested lists, such as
\strong{glance} and \strong{tidy} and flattens the table into one level, such that the
nested tables are expanded out. This is essentially a wrapper for
\code{\link[tidyr:nest]{tidyr::unnest()}}.
}
\seealso{
Other extractors: 
\code{\link{extract_results}()}
}
\concept{extractors}
