% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{fmlss}
\alias{fmlss}
\alias{sx}
\alias{cast_fmls}
\title{Vectorized formulas}
\usage{
sx(
  x = unspecified(),
  role = list(),
  tier = list(),
  label = list(),
  pattern = character(),
  ...
)

cast_fmls(
  x = unspecified(),
  role = list(),
  tier = list(),
  label = list(),
  pattern = character(),
  ...
)
}
\arguments{
\item{x}{Objects of the following types can be used as inputs
\itemize{
\item \code{tm}
\item \code{formula}
}}

\item{pattern}{This is the expansion pattern used to decide how the
covariates will incorporated into the formulas. The options are
\code{c("direct", "sequential", "parallel")}. See the details for further
explanation.
\itemize{
\item \strong{direct}: the covariates will all be included in each formula
\item \strong{sequential}: the covariates will be added sequentially, one by one, or
by tiers, as indicated
\item \strong{parallel}: the covariates or tiers of covariates will be placed in
parallel
}}

\item{...}{Arguments to be passed to or from other methods}
}
\value{
An object of class \code{fmls}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function defines a modified \code{formula} class that has been
vectorized. The \code{fmls} serves as a set of instructions or a \emph{script} for the
formula and its tm. It expands upon the functionality of formulas,
allowing for additional descriptions and relationships to exist between the
tm.
}
\section{Roles}{


Specific roles the variable plays within the formula. These are of particular
importance, as they serve as special tm that can effect how a formula is
interpreted. The specialized options for roles are as below:
\itemize{
\item \strong{exposure} or \code{X(...)}
\item \strong{outcome} or \code{O(...)} or placement of variable on LHS of formula
\item \strong{confounder} or \code{C(...)}
\item \strong{mediator} or \code{M(...)}
\item \strong{strata} or \code{S(...)}
\item \strong{interaction} or \code{In()}
}

Formulas can be condensed by applying their specific role to individual tm
as a function/wrapper. For example, \code{y ~ X(x1) + x2 + x3}. This would signify
that \code{x1} has the specific role of an exposure.
}

\section{Patterns}{


The expansion pattern allows for instructions on how the covariates should be
included in different formulas. Below, assuming that \emph{x1}, \emph{x2}, and \emph{x3} are
covariates...

\deqn{y ~ x1 + x2 + x3}

\strong{Direct}:

\deqn{y ~ x1 + x2 + x3}

\strong{Seqential}:

\deqn{y ~ x1}
\deqn{y ~ x1 + x2}
\deqn{y ~ x1 + x2 + x3}

\strong{Parallel}:

\deqn{y ~ x1}
\deqn{y ~ x2}
\deqn{y ~ x3}
}

