% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forge.R
\name{forge}
\alias{forge}
\alias{forge.model_archetype}
\alias{forge.formula_archetype}
\alias{mdls}
\title{Forge Models into a Table}
\usage{
forge(x, ...)

\method{forge}{model_archetype}(x, name = deparse1(substitute(x)), data = data.frame(), ...)

\method{forge}{formula_archetype}(x, name = deparse1(substitute(x)), data = data.frame(), ...)

mdls(x, ...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function introduces a super class that combines both the \code{list} class
(and its derivative \code{list_of}) and regression models and/or hypothesis tests.
Models that are similar and share certain properties can be combined together
into a \code{forge}.
}
