% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{formulas_to_named_list}
\alias{formulas_to_named_list}
\title{Take list of formula, and return as a named list (name = LHS, value = RHS)}
\usage{
formulas_to_named_list(x)
}
\description{
Take list of formula, and return as a named list (name = LHS, value = RHS)
}
