% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framework.R
\name{framework}
\alias{framework}
\title{Initialize a Hypothesis Framework}
\usage{
framework(...)
}
\description{
Initialize a Hypothesis Framework
}
\seealso{
Other frameworks: 
\code{\link{add_hypothesis}()}
}
\concept{frameworks}
