% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesize.R
\name{hypothesize}
\alias{hypothesize}
\alias{hypothesize.default}
\alias{hypothesize.formula}
\alias{hypothesize.list}
\alias{hypothesize.dagitty}
\alias{hypothesize.tidy_dagitty}
\alias{hypothesize.recipe}
\alias{hypothesize.rx}
\title{Draw A Hypothesis}
\usage{
hypothesize(x, ...)

\method{hypothesize}{formula}(
  x,
  exposures = NULL,
  fixed = NULL,
  combination,
  test,
  test_opts = NA,
  data,
  strata = NA,
  origin = "independent",
  ...
)

\method{hypothesize}{rx}(
  x,
  combination,
  test,
  test_opts = NA,
  data,
  strata = NA,
  origin = "independent",
  ...
)

\method{hypothesize}{list}(x, ...)

\method{hypothesize}{dagitty}(x, ...)

\method{hypothesize}{tidy_dagitty}(x, ...)

\method{hypothesize}{recipe}(h, ...)

\method{hypothesize}{default}(x, ...)
}
\arguments{
\item{x}{General variable that can be cast into a hypothesis object.
Currently this accepts the following object types:
\itemize{
\item \code{formula} objects, which can be given additional modifying label. The key
variables, as described in the \strong{formulae section} below, are \code{exposures}
and \code{fixed}, as they allow for specific transformations of the formula.
This behavior is also described in \code{\link[=expand_formula]{expand_formula()}}.
\item \code{rx} objects, which are modified formulas created by \code{\link[rx:rx-package]{rx::rx()}} and have
additional labels available. Can be set by \code{\link[rx:options]{rx::set_rx_theme()}} using the
argument \code{set_rx_theme("epi")}. This behavior is described in
\code{\link[=expand_formula]{expand_formula()}} or the formula section below with further details.
\item \code{list} object, which describes the outcomes, exposures, covariates, fixed
variables, and potential confounders in a named list, which allows for
grouped covariates when modeling. Please see the details section for
further explanation.
}}

\item{...}{For additional variables based on the generic method invoked.}

\item{exposures}{A character vector for individual exposures, which will
always be placed in separate models.}

\item{fixed}{A character vector naming variables that should be retained in
all models.}

\item{combination}{The building pattern for how to put together the overall
plan. It defines variable relationships that will be used. The options for
the \code{combination} currently include:
\itemize{
\item \code{direct} will define the relationship as y ~ x
\item \code{sequential} will define the relationship as y ~ x1, y ~ x1 + x2
\item \code{parallel} will define the relationship as y ~ x1, y ~ x2
}}

\item{test}{The testing that will be used, defined as how the analysis or
hypothesis testing should be performed. There are several options on which
specification to use, anywhere from regression modeling to inferential
statistics. The test applies to the collection of hypotheses.
\itemize{
\item A model specified by the \href{https://parsnip.tidymodels.org}{\code{parsnip} package}, which includes the mode and
computational engine
\item A statistical test, such as a \code{t.test}, which may require additional
parameters. These can be given as additional, unmatched arguments. This
option currently supports only hypothesis tests, of class \code{htest}.
}}

\item{test_opts}{For optional or additional parameters to be given to the
\emph{test} argument. Should be in format of a list with name-pairs, such as
\code{list(name = argument)} pairs. Default is NA, indicating there are no test
options to pass along.}

\item{data}{Data set to be linked to this hypothesis. It assumes data is one
of type \code{c("tbl_df", "tbl", "data.frame")}. The name of the data set is
taken from the object itself.}

\item{strata}{How the data should be split or stratified. References the
name of the data given in that the models will be with fit against,
splitting the data into subsets. This helps to perform hypothesis testing
on subsets or strata of the data. It defaults to \strong{NA} (which means the
full data will be used). This argument will only be incorporated if \code{data}
is present.}

\item{origin}{Variable that serves as an attribute to help track the
hypothesis evolution through the study.}
}
\value{
An object of class \code{hypothesis}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{hypothesis()} creates a \code{hypothesis} object that can be placed into a
modeling framework. This framework allows for multiple hypotheses to be
modeled, and allows for assessment of confounding variables.
}
\section{Warning}{


The \code{hypothesis} object does not yet fully work for function-based
\strong{tests}, such as:
\itemize{
\item \code{htest} objects from functions like \code{\link[stats:t.test]{stats::t.test()}} and
\code{\link[stats:cor.test]{stats::cor.test()}}
\item \code{glm} and \code{lm} objects from functions like \code{\link[stats:lm]{stats::lm()}} and
\code{\link[stats:glm]{stats::glm()}}
}

The \code{hypothesis} doesn't have full functionality for the follow objects
yet:
\itemize{
\item \code{dagitty} from \code{\link[dagitty:dagitty]{dagitty::dagitty()}} or \code{\link[ggdag:dagify]{ggdag::dagify()}}
\item \code{tidy_dagitty} from \code{\link[ggdag:tidy_dagitty]{ggdag::tidy_dagitty()}}
\item \code{recipe} from \code{\link[recipes:recipe]{recipes::recipe()}}
\item \code{list} objects
}
}

\section{Formulae}{


For each \code{formula}, the terms in the RHS and LHS can be labeled. Each individual LHS term is considered an individual outcome, and each RHS is a predictor which can be grouped in a variety of ways. If labels are provided, then the \code{formula} will be expanded in a certain pattern. The key labels are:
\itemize{
\item \strong{exposures} = defines an independent exposure, which will be placed in a separate formulae from any other term marked as such
\item \strong{fixed} = defines any predictor that should be maintained/fixed in all models, including complex terms such as mixed/random effects
}
}

\concept{hypotheses}
