% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms.R
\name{intersect.tm}
\alias{intersect.tm}
\title{Set operations for \code{tm} objects}
\usage{
\method{intersect}{tm}(x, y, ...)
}
\arguments{
\item{x, y}{Pair of \code{tm} vectors}

\item{...}{These dots are for future extensions and must be empty}
}
\description{
Set operations for \code{tm} objects
}
