% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framework-helpers.R
\name{link_framework}
\alias{link_framework}
\alias{.link_hypothesis}
\alias{.link_test}
\alias{.link_data}
\alias{.link_status}
\title{Linking Functions Between Hypothesis and Framework}
\usage{
.link_hypothesis(framework, hypothesis, name)

.link_test(framework, hypothesis, name)

.link_data(framework, hypothesis, name)

.link_status(framework, hypothesis, name)
}
\arguments{
\item{framework}{Object of \code{framework} class}

\item{hypothesis}{Object of \code{hypothesis} class}

\item{name}{Name of \code{hypothesis}, given from parent function}
}
\description{
These are all internal functions to help attach components to the \code{framework} object, and are generally not exposed to the user.
}
\keyword{internal}
