% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study-helpers.R
\name{link_study}
\alias{link_study}
\alias{.link_hypothesis}
\alias{.link_test}
\alias{.link_data}
\alias{.link_status}
\alias{.link_vars}
\title{Linking Functions Between Hypothesis and Study}
\usage{
.link_hypothesis(study, hypothesis, name)

.link_test(study, hypothesis, name)

.link_data(study, hypothesis, name)

.link_status(study, hypothesis, name)

.link_vars(study, hypothesis, name)
}
\arguments{
\item{study}{Object of \code{study} class}

\item{hypothesis}{Object of \code{hypothesis} class}

\item{name}{Name of \code{hypothesis}, given from parent function}
}
\description{
These are all internal functions to help attach components to the \code{study}
object, and are generally not exposed to the user.
}
\keyword{internal}
