% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-of-formulas.R
\name{list_of_formulas}
\alias{list_of_formulas}
\alias{lst_fmls}
\title{List of formulas}
\usage{
lst_fmls(...)

list_of_formulas(...)
}
\description{
Vectorized subclass of the \code{list_of} class from \code{\link[vctrs:list_of]{vctrs::list_of()}} that
utilizes a formula pattern for containing information or relationships. This
serves as a base class that can be expanded as needed.
}
