% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_hypothesis.R
\name{make_formulas}
\alias{make_formulas}
\title{Make formula combinations in \code{add_hypothesis()}}
\usage{
make_formulas(formula, combination)
}
\arguments{
\item{formula}{Formula showing relationship of outcomes and predictors, and
is essentially the hypothesis. It allows for complex formulas e.g. multiple
predictors, multiple exposures, and multiple covariates, that can be
organized into individual formulas based on additional modifiers. Each RHS
term of the formula is considered an outcome variable, and is analyzed as a
single outcome. Each LHS term is consider a predictor, and can be modified
as below:
\itemize{
\item \code{exp()} is placed around a term to define as an independent exposure,
which will be placed in separate formulas from any other term marked as an
exposure
\item \code{cov()} is placed around a term for any predictors that should be
maintained/fixed in all models, which can include complex terms, such as mixed
effects
}

Of note, these formula modifiers do not apply to \code{htest} objects yet.

For example, the equation below describes two independent exposures "x1" and "x2" that should be conditionally regressed for every level of "z"

\deqn{y ~ exp(x1) + exp(x2) + x3 + x4 + cov((1 | z))}}

\item{combination}{The building pattern for how to put together the overall
plan. It defines variable relationships that will be used. The options for
the \code{combination} currently include:
\itemize{
\item \code{direct} will define the relationship as y ~ x
\item \code{sequential} will define the relationship as y ~ x1, y ~ x1 + x2
\item \code{parallel} will define the relationship as y ~ x1, y ~ x2
}}
}
\description{
This is an internal function that helps to create the appropriate formula combinations for adding hypotheses.
}
