% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{map_dag}
\alias{map_dag}
\title{Map Out A Directed Acyclic Graph}
\usage{
map_dag(study, name, tidy = FALSE)
}
\arguments{
\item{study}{A \code{study} object}

\item{name}{The name of a hypothesis added to the study}

\item{tidy}{Defaults to FALSE, thus returning a \code{dagitty} object. If TRUE, then will return a \code{tidy_dagitty} object.}
}
\value{
\code{dagitty} or \code{tidy_dagitty} object
}
\description{
This function converts a hypothesis into a \code{dagitty} object (or
\code{tidy_dagitty} if requested). This can subsequently be passed onto the
\code{\link[ggdag:ggdag]{ggdag::ggdag()}} function for additional plotting.
}
