% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-table.R
\name{md_tbl}
\alias{md_tbl}
\alias{model_table}
\title{Many models in a table}
\usage{
md_tbl(..., data = NULL)

model_table(..., data = NULL)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function introduces a super class that combines both the \code{list} class
(and its derivative \code{list_of}) and regression models and/or hypothesis tests.
Models that are similar and share certain properties can be combined together
into a \code{md_tbl}.
}
