% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_lists.R
\name{model_list}
\alias{model_list}
\alias{model_list.list}
\alias{model_list.default}
\alias{mdls}
\title{Model Lists}
\usage{
model_list(x, ...)

\method{model_list}{list}(
  x,
  labels = list(),
  roles = list(),
  name = deparse1(substitute(x)),
  ...
)

\method{model_list}{default}(x, ...)

mdls(x, ...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function introduces a super  that combines both the \code{list} class (and
its derivative \code{list_of}) and regression models
}
