% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suits.R
\name{model_suit}
\alias{model_suit}
\alias{model_suit.formula_list}
\alias{model_suit.default}
\alias{mdls}
\title{Model Lists}
\usage{
model_suit(x, ...)

\method{model_suit}{formula_list}(x, name = deparse1(substitute(x)), fitting_function, ..., data)

\method{model_suit}{default}(x, ...)

mdls(x, ...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function introduces a super class that combines both the \code{list} class
(and its derivative \code{list_of}) and regression models and/or hypothesis tests.
Models that are similar and share certain properties can be combined together
into a \code{model_suit}.
}
