% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forge.R
\name{model_table}
\alias{model_table}
\alias{md_tbl}
\title{Forge Models into a Table}
\usage{
model_table(x, ...)

md_tbl(x, ...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function introduces a super class that combines both the \code{list} class
(and its derivative \code{list_of}) and regression models and/or hypothesis tests.
Models that are similar and share certain properties can be combined together
into a \code{model_table}.
}
