% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_hypothesis.R, R/maps.R
\name{modify_map}
\alias{modify_map}
\alias{.integrate_formula}
\alias{.integrate_data}
\alias{.integrate_test}
\alias{.integrate_status}
\alias{.revise_strata}
\alias{.revise_status}
\title{Hypothesis Decomposers to Integrate into Model Map}
\usage{
.integrate_formula(model_map, hypothesis, name)

.integrate_data(model_map, hypothesis, name)

.integrate_test(model_map, hypothesis, name)

.integrate_status(
  model_map,
  hypothesis,
  name,
  run = FALSE,
  tidy = FALSE,
  error = FALSE,
  origin = NA
)

.revise_strata(model_map, hypothesis, name)

.revise_status(model_map, name, ...)
}
\arguments{
\item{model_map}{Object of \code{model_map} class}

\item{hypothesis}{Object of \code{hypothesis} class}

\item{name}{Name of the \code{hypothesis} object, which defaults to the name of
the \code{hypothesis} object itself}

\item{...}{For extensibility}
}
\value{
A \code{model_map} object
}
\description{
Break apart a \code{hypothesis} and add its components into the appropriate part of a \code{model_map} object.
}
\concept{internals}
\keyword{internal}
