% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study-helpers.R
\name{modify_study}
\alias{modify_study}
\alias{add_study_formula}
\alias{add_study_paths}
\alias{add_study_test}
\alias{add_study_data}
\alias{add_study_status}
\title{Study Modifications}
\usage{
add_study_formula(study, hypothesis, name)

add_study_paths(study, hypothesis, name)

add_study_test(study, hypothesis, name)

add_study_data(study, hypothesis, name)

add_study_status(
  study,
  hypothesis,
  name,
  run = FALSE,
  error = FALSE,
  stage = NA,
  path = FALSE
)
}
\arguments{
\item{study}{Object of \code{study} class}

\item{hypothesis}{Object of \code{hypothesis} class}

\item{name}{Name of \code{hypothesis}, given from parent function}
}
\description{
Study Modifications
}
\keyword{internal}
