% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{mx}
\alias{mx}
\alias{model_rx}
\alias{model_rx.lm}
\alias{model_rx.default}
\title{Model records}
\usage{
model_rx(x, ...)

\method{model_rx}{lm}(
  x,
  model_name = deparse(substitute(x)),
  term_labels = list(),
  term_roles = list(),
  model_label = character(),
  model_description = character(),
  ...
)

\method{model_rx}{default}(x, ...)

mx(x, ...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
