% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-table.R
\name{new_model_table}
\alias{new_model_table}
\title{Initialize new model-based tibble / tbl_df}
\usage{
new_model_table(
  x = list(),
  formulaMatrix = data.frame(),
  termTable = data.frame(),
  ...
)
}
\description{
Initialize new model-based tibble / tbl_df
}
\keyword{internal}
