% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equip.R
\name{outfit}
\alias{outfit}
\alias{equip}
\title{Equip the \code{octomod} With Hypothesis Testing}
\usage{
equip(octomod, which_arms = NULL, ...)
}
\arguments{
\item{octomod}{Object of class \code{octomod}}

\item{which_arms}{Vector of names of arms that should be run. Defaults to all
arms that have not yet been equipped..}

\item{...}{To pass additional parameters as needed}
}
\value{
An \code{octomod} object with arms outfitted with the test statistics.
}
\description{
This function allows for fitting the many arms of the \code{octomod} object,
equipping them with the correct test methods. For modeling, this heavily
leans on the \code{tidymodels} approach which allows for unified model
specification. For inferential statistics, which do not yet have a unified
testing approach, this provides the ability to run these somewhat simpler
tests in a cohesive manner.
}
\examples{
library(magrittr)
library(parsnip)
lm_mod <- linear_reg() \%>\% set_engine("lm")

om <-
  octomod() \%>\%
  core(iris) \%>\%
  arm(
    title = "t_test",
    plan = Sepal.Length + Sepal.Width ~ Petal.Length,
    pattern = "direct",
    approach = "t.test",
    paired = TRUE
  ) \%>\%
  arm(
    title = "linear",
    plan = Petal.Width ~ Sepal.Length + Sepal.Width + Petal.Length,
    pattern = "sequential",
    approach = lm_mod
  ) \%>\%
  equip()

}
