% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{pattern}
\alias{pattern}
\alias{pattern_sequentially}
\title{Expand formula by a pattern
Must occur after the formula has been simplified}
\usage{
pattern(x, ...)

pattern_sequentially(x, ...)
}
\value{
\code{fmls} objects expanded by pattern
}
\description{
Expand formula by a pattern
Must occur after the formula has been simplified
}
