% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{patterns}
\alias{patterns}
\alias{pattern}
\alias{pattern_sequential}
\alias{pattern_parallel}
\alias{pattern_direct}
\alias{pattern_fundamental}
\title{Expand formula by a pattern
Must occur after the formula has been simplified}
\usage{
pattern(x, ...)

pattern_sequential(x, ...)

pattern_parallel(x, ...)

pattern_direct(x, ...)

pattern_fundamental(x, ...)
}
\value{
\code{fmls} objects expanded by pattern
}
\description{
Expand formula by a pattern
Must occur after the formula has been simplified
}
