% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_encoded_vimp.R
\name{plot_encoded_vimp}
\alias{plot_encoded_vimp}
\title{Plot Variable Importance for Encoded Dimensions}
\usage{
plot_encoded_vimp(
  encoded_data,
  plot_colors = c("steelblue", "red"),
  flip = FALSE
)
}
\arguments{
\item{encoded_data}{A data frame containing the encoded dimension importance data, created from the
\code{vimp_encoded} function. It should have columns: Class, EncodedDimension, and Importance.}

\item{plot_colors}{A vector of two colors for the gradient fill.
Default is c("steelblue", "red").}

\item{flip}{If TRUE, then flip the bar plot to be displayed vertically.}
}
\value{
A ggplot object of a facet bar plot for each class showing the importance for each encoded dimension.
}
\description{
This function creates a ggplot visualisation of the variable importance
for encoded dimensions, typically output from the \code{vimp_encoded} function.
}
\details{
The function creates a faceted bar plot where each facet represents a class.
The bars represent the importance of each encoded dimension, with colour
intensity indicating the level of importance.
}
