% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_input.R
\name{plot_input}
\alias{plot_input}
\title{Plot Feature Importance for Each Dimension}
\usage{
plot_input(
  input_vimp,
  sort = FALSE,
  colors = c("white", "red"),
  flip_horizontal = FALSE,
  flip_vertical = FALSE,
  rotate = 0
)
}
\arguments{
\item{input_vimp}{A list obtained from the \code{vimp_input} function.}

\item{sort}{Logical. If TRUE, sorts the input importance plots based on their value range. Default is FALSE.}

\item{colors}{A vector of two colours for the gradient scale. Default is c("white", "red").}

\item{flip_horizontal}{Logical. If TRUE, flips the plot horizontally. Default is FALSE.}

\item{flip_vertical}{Logical. If TRUE, flips the plot vertically. Default is FALSE.}

\item{rotate}{Integer. Specifies the rotation of the plot. 0 = no rotation, 1 = 90 degrees, 2 = 180 degrees, 3 = 270 degrees.
Default is 0.}
}
\value{
A grid of plots, each representing the feature importance for one
        dimension of the encoded data.
}
\description{
Creates a series of plots, displayed on a grid,
showing the feature importance for each dimension of the encoded data.
}
\details{
The function generates a grid of plots, one for each
dimension of the encoded data, where each plot visualises the feature importance
as a heatmap.
}
