% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_input_direction.R
\name{plot_input_direction}
\alias{plot_input_direction}
\title{Plot Importance Direction}
\usage{
plot_input_direction(
  input_vimp,
  encoded_vimp,
  direction_vimp,
  class,
  topX = NULL,
  sort = FALSE,
  filter_zero = FALSE,
  flip_horizontal = FALSE,
  flip_vertical = FALSE,
  rotate = 0
)
}
\arguments{
\item{input_vimp}{A list containing input variable importance data. Created via the \code{vimp_input} function.}

\item{encoded_vimp}{A data frame with encoded dimension variable importance data. Created via the \code{vimp_encoded} function.}

\item{direction_vimp}{A matrix with linear model summary data. Created via the \code{vimp_direction} function.}

\item{class}{The specific class or classes to analyse.}

\item{topX}{Integer. If provided, limits the visualisation to the top X most important dimensions.}

\item{sort}{Logical. If TRUE, sorts the importance values. Default is FALSE.}

\item{filter_zero}{Logical. If TRUE, filters out zero importance values (i.e., unused dimensions). Default is FALSE.}

\item{flip_horizontal}{Logical. If TRUE, flips the heatmaps horizontally. Default is FALSE.}

\item{flip_vertical}{Logical. If TRUE, flips the heatmaps vertically. Default is FALSE.}

\item{rotate}{Integer. Specifies the rotation of the heatmaps. 0 = no rotation, 1 = 90 degrees, 2 = 180 degrees, 3 = 270 degrees. Default is 0.}
}
\value{
A grid arrangement of plots showing encoded dimension importance and pixel importance direction.
}
\description{
This function creates visualisations of the encoded dimension importance and the input importance
with the direction of the importance highlighted.
It generates a bar plot of encoded dimension importance and heatmaps showing pixel importance direction.
}
