% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAvgDecoded.R
\name{plotAvgDecoded}
\alias{plotAvgDecoded}
\title{Plot Average Decoded Values}
\usage{
plotAvgDecoded(data, colours = c("white", "red"), legendName = "Avg Error")
}
\arguments{
\item{data}{A data frame created from \code{vimpDecoded} function.}

\item{colours}{A character vector of length 2 specifying the gradient colors for the heatmap.
Defaults to \code{c('white', 'red')}. The first element is used for low values, and the second for high values.}

\item{legendName}{The name to be used for the heatmap legend. Default is 'Avg Error'.}
}
\value{
A ggplot object.
}
\description{
Creates a bar plot visualization showing the average decoded values over all selected digits
}
\details{
The function calculates the average importance across all dimensions.
}
