% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAvgEncoded.R
\name{plotAvgEncoded}
\alias{plotAvgEncoded}
\title{Plot Average Encoded Values}
\usage{
plotAvgEncoded(
  digitResults,
  colours = c("white", "red"),
  legendName = "Avg Error"
)
}
\arguments{
\item{digitResults}{A list containing the selected digit/s and its/their corresponding importance values.}

\item{colours}{A character vector of length 2 specifying the gradient colors for the heatmap.
Defaults to \code{c('white', 'red')}. The first element is used for low values, and the second for high values.}

\item{legendName}{The name to be used for the heatmap legend. Default is 'Avg Error'.}
}
\value{
A ggplot object representing the heatmap.
}
\description{
Creates a heatmap visualization showing the average encoded values for a given digit.
}
\details{
The function calculates the average importance across all dimensions and creates a heatmap.
The heatmap represents the average permutation importance for the selected digit,
with each pixel's importance depicted.
}
\examples{
\dontrun{
# Train focusing only on the digit 1 for 10 epochs
encoderModel <- mnistAuto(numSel = 1, epochs = 10)
vimps <- vimpEncoded(encoderModel)
plotAvgEncoded(digitResults = vimps)
}
}
