% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEncoded.R
\name{plotEncoded}
\alias{plotEncoded}
\title{Plot Feature Importance for Each Dimension}
\usage{
plotEncoded(digitResults, legendName = "Error", colours = c("white", "red"))
}
\arguments{
\item{digitResults}{A list obtained from the \code{vimpEncoded} function.}

\item{legendName}{The name to be used for the heatmap legend. Default is 'Error'.}

\item{colours}{The selected colours to display the heatmap. First value corresponds to
low values}
}
\value{
A grid of plots, each representing the feature importance for one
        dimension of the encoded data.
}
\description{
Creates a series of plots, displayed on a grid,
showing the feature importance for each dimension of the encoded data.
}
\details{
The function generates a grid of plots, one for each
dimension of the encoded data, where each plot visualizes the feature importance
as a heatmap.
}
\examples{
# Train focusing only on the digit 1 for 10 epochs
encoderModel <- mnistAuto(numSel = 1, epochs = 10)
vimps <- vimpEncoded(encoderModel)
plotEncoded(digitResults = vimps)

}
