% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEncodedSpace.R
\name{plotEncodedSpace}
\alias{plotEncodedSpace}
\title{Plot Encoded Space of Encoder Model}
\usage{
plotEncodedSpace(
  encoderModel,
  type = "std",
  perplexity = 30,
  tsneDims = 2,
  verbose = TRUE
)
}
\arguments{
\item{encoderModel}{A list containing the encoder model at the first position and the test data x_test at the second position.}

\item{type}{Character string specifying the type of plot. 'std' for standard deviation-based plot and 'tsne' for t-SNE based plot. Defaults to 'std'.}

\item{perplexity}{Numeric. The perplexity parameter used in the t-SNE algorithm. Typically values between 5 and 50 are used. The default is 30.}

\item{tsneDims}{Integer. The number of dimensions for the t-SNE reduced space. Can be either 2 or 3. The default is 2.}

\item{verbose}{Logical. If TRUE, additional information will be printed during the execution of t-SNE algorithm. Defaults to TRUE.}
}
\value{
A ggplot object representing the encoded space.
}
\description{
This function takes an encoder model and its associated test data to plot the encoded space.
It supports both standard deviation-based plotting and t-SNE (t-Distributed Stochastic Neighbor Embedding) for dimensionality reduction.
}
\examples{
\dontrun{
plot <- plotEncodedSpace(encoderModel, type = 'std')
plot(plot) # To display the plot

# For a t-SNE based plot with specific perplexity
tsne_plot <- plotEncodedSpace(encoderModel, type = 'tsne', perplexity = 40)
plot(tsne_plot)
}
}
