% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{print.model_map}
\alias{print.model_map}
\title{Print a Model Map}
\usage{
\method{print}{model_map}(x, ...)
}
\arguments{
\item{x}{A \code{model_map} object}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print a Model Map
}
