% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{print.study}
\alias{print.study}
\title{Print a Study}
\usage{
\method{print}{study}(x, ...)
}
\arguments{
\item{x}{A \code{study} object}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print a Study
}
