% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project}
\alias{project}
\title{Initialize a Project}
\usage{
project()
}
\value{
\code{project} object
}
\description{
This function creates the base structure for a research project. Specific
tests and models can be added, and subsequently built, upon a prescribed data
set.
}
