% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confound.R
\name{reconstruct}
\alias{reconstruct}
\title{Reconstruct a Hypothesis}
\usage{
reconstruct(
  model_map,
  name,
  new_name = paste0(name, "_cut"),
  approach = "confounding",
  ...
)
}
\arguments{
\item{model_map}{A \code{model_map} object that has been constructed}

\item{name}{Name of a an existing \code{hypothesis} object}

\item{new_name}{Name of the new \code{hypothesis} that is created after
modification. Defaults to modifying the original \strong{name} by appending
\verb{*_cut}.}

\item{approach}{String that informs \emph{how} to reconstruct the
hypothesis. Options include:
\itemize{
\item \strong{confounding} = Check for relevant terms based on
\code{\link[dagger:find_confounders]{dagger::find_confounders()}}. This includes the \strong{delta} argument to
set the threshold for changes in effect size that are considered relevant.
}}

\item{...}{Additional, optional parameters based on approach being used}
}
\value{
A \code{model_map} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Modify a \code{hypothesis} within a \code{model_map} with different approaches.
}
\concept{confounders studies}
