% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confound.R
\name{reconstruct_map}
\alias{reconstruct_map}
\title{Reconstruct a Hypothesis}
\usage{
reconstruct_map(
  study,
  name,
  new_name = paste0(name, "_cut"),
  approach = "confounding",
  ...
)
}
\arguments{
\item{study}{A \code{study} object that has been constructed}

\item{name}{Name of a an existing \code{hypothesis} object}

\item{new_name}{Name of the new \code{hypothesis} that is created after
modification. Defaults to modifying the original \strong{name} by appending
\verb{*_cut}.}

\item{approach}{String that informs \emph{how} to reconstruct the
hypothesis. Options include:
\itemize{
\item \strong{confounding} = Check for relevant terms based on [\code{\link[=find_confounders]{find_confounders()}}]
}}

\item{...}{Additional, optional parameters based on approach being used}
}
\value{
A \code{study} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Modify a \code{hypothesis} within a study with different approaches.
}
