% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{return_cols}
\alias{return_cols}
\title{Return name of column from either \emph{character} or \emph{formula ~ list} pattern}
\usage{
return_cols(x)
}
\description{
Return name of column from either \emph{character} or \emph{formula ~ list} pattern
}
