% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{return_names}
\alias{return_names}
\title{Return the names from either \emph{character} or \emph{formula ~ list} pattern}
\usage{
return_names(x)
}
\description{
Return the names from either \emph{character} or \emph{formula ~ list} pattern
}
