% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{return_rows}
\alias{return_rows}
\title{Return important rows from either \emph{character} or \emph{formula ~ list} pattern}
\usage{
return_rows(x)
}
\description{
Return important rows from either \emph{character} or \emph{formula ~ list} pattern
}
