% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdl-package.R
\docType{package}
\name{rmdl-package}
\alias{rmdl}
\alias{rmdl-package}
\title{rmdl package}
\description{
The language provided is intended to help organize many models that are generated in the process of causal inference in data analysis. The specific classes and functions emphasize the importance of variable relationships (e.g. exposure, outcome, confounding terms) and helps contextualize formulas and models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/asshah4/rmdl}
  \item Report bugs at \url{https://github.com/asshah4/rmdl/issues}
}

}
\author{
\strong{Maintainer}: Anish Shah \email{ashah282@uic.edu} (\href{https://orcid.org/0000-0002-9729-1558}{ORCID}) [copyright holder]

}
\keyword{internal}
