% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms.R
\name{roles}
\alias{roles}
\title{Set operations for \code{tm} objects}
\usage{
roles(x)
}
\arguments{
\item{x, y}{Pair of \code{tm} vectors}

\item{...}{These dots are for future extensions and must be empty
Retrieve components of \code{tm} objects

Family of related functions that extract components of \code{tm} objects.}

\item{x}{A vector \code{tm} objects}
}
\description{
Set operations for \code{tm} objects
}
