% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_data.R
\name{set_data}
\alias{set_data}
\title{Set the Data}
\usage{
set_data(project, .data, .stage = "data", ...)
}
\arguments{
\item{project}{Object of class \code{project}}

\item{.data}{Data frame or tibble}

\item{.stage}{Internal marker of workflow progress.}

\item{...}{For extensibility}
}
\value{
\code{project} object
}
\description{
Define the data set that should be used for analysis or tests. The data will
be paired with the hypotheses or tests that will be performed. If the project
already has a data set added, another will be added without effecting prior
tests.
}
