% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis-helpers.R
\name{set_hypothesis}
\alias{set_hypothesis}
\alias{.set_formulae}
\alias{.set_data}
\alias{.set_tests}
\title{Set Components of a Hypothesis}
\usage{
.set_formulae(hypothesis, formula, combination)

.set_data(hypothesis, data, data_name, strata = NULL)

.set_tests(hypothesis, test, .test_opts)
}
\arguments{
\item{hypothesis}{A \code{hypothesis} object}

\item{combination}{The pattern of how to generate formulas}

\item{data}{A data frame object}

\item{data_name}{Abstracted from the \strong{data} argument}

\item{strata}{Optional argument to subset data frame}
}
\value{
Invisibly returns the \code{hypothesis} object
}
\description{
The following functions will replace the internal attributes of a
\code{hypothesis} object, without warning.
\itemize{
\item \code{.set_formulae()} expands the formula based on outcomes and predictors
(including both exposures, covariates, and fixed variables). The pattern is
decided by the \code{combination} argument. The major functionality is within the
\code{expand_formula()} function, but this allows for setting internal attributes.
\item \code{.set_data()} adds data to a formula (internal to the \code{hypothesize()}
function), along with potential stratification instructions. Modifies the
formula by adding the attributes to the formula.
\item \code{set_tests()} Add the test and test options (if needed) the the
\code{hypothesis} object. This modifies the attributes of the original object and
invisibly returns the same object.
}
}
\keyword{internal}
