% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharpen.R
\name{sharpen}
\alias{sharpen}
\title{Sharpen the Hypothesis Testing with Statistical Metrics #' #' This function adds a new column to any \code{arm} that has been passed to #' \code{equip()}. It only applies to \code{parsnip} models currently, as defined in the #' \strong{approach} from the respective arm. It provides a test statistic on model #' fit for the original \code{core} data. This new column is named \strong{metric}. It is #' intended to expand from the \strong{tidied} column to include information about #' the overall model. See \strong{details} for currently supported tests. #' #' It currently supports: #' #'   - \code{logistic_reg} = The function will return \emph{only} the models that are #'   generated by \code{\link[parsnip:logistic_reg]{parsnip::logistic_reg()}}. It generates a simple c-statistic
(or AUC from ROC curve). The default for this is that the binary event has
levels \code{c(0, 1)}, with the event of interest being \code{1} (e.g. the second
event).}
\usage{
sharpen(octomod, which_arms = NULL, ...)
}
\arguments{
\item{octomod}{Object of class \code{octomod}}

\item{which_arms}{Vector of names of arms that should be run. Defaults to all
arms that have not yet been equipped..}

\item{...}{To pass additional parameters as needed}
}
\value{
An \code{octomod} object with arms outfitted with the test statistics.
}
\description{
Sharpen the Hypothesis Testing with Statistical Metrics #' #' This function adds a new column to any \code{arm} that has been passed to #' \code{equip()}. It only applies to \code{parsnip} models currently, as defined in the #' \strong{approach} from the respective arm. It provides a test statistic on model #' fit for the original \code{core} data. This new column is named \strong{metric}. It is #' intended to expand from the \strong{tidied} column to include information about #' the overall model. See \strong{details} for currently supported tests. #' #' It currently supports: #' #'   - \code{logistic_reg} = The function will return \emph{only} the models that are #'   generated by \code{\link[parsnip:logistic_reg]{parsnip::logistic_reg()}}. It generates a simple c-statistic
(or AUC from ROC curve). The default for this is that the binary event has
levels \code{c(0, 1)}, with the event of interest being \code{1} (e.g. the second
event).
}
\examples{
library(magrittr)
library(parsnip)
df <- mtcars
df$am <- factor(df$am)
df$vs <- factor(df$vs)

om <-
  octomod() \%>\%
  core(df) \%>\%
  arm(
    title = "log_model",
    f = am + vs ~ mpg + disp + hp,
    pattern = "parallel",
    approach = logistic_reg() \%>\% set_engine("glm"),
  ) \%>\%
  equip() \%>\%
  sharpen(which_arms = "log_model")

}
