% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_vimp.R
\name{shiny_vimp}
\alias{shiny_vimp}
\title{Launch Shiny App for Autoencoder Analysis Visualisation}
\usage{
shiny_vimp(input_vimp, encoded_vimp, direction_vimp)
}
\arguments{
\item{input_vimp}{A list or data frame containing input variable importance data. Created via the \code{vimp_input} function.}

\item{encoded_vimp}{A list or data frame containing encoded variable importance data. Created via the \code{vimp_encoded} function.}

\item{direction_vimp}{A list or data frame containing direction variable importance data. Created via the \code{vimp_direction} function.}
}
\value{
A Shiny app object.
}
\description{
This function creates and launches a Shiny application for interactive visualisation
of autoencoder analysis results, including variable importance for both the
input pixels and the encoded dimension.
}
\details{
The app provides interactive controls for:
\itemize{
  \item Selecting a specific class to analyse
  \item Choosing the rotation of the plots
  \item Specifying the number of top importances to display
  \item Filter Empty Dimensions
  \item Sorting by importance
  \item Flipping the plots horizontally or vertically
  \item Switching between light and dark mode themes.
}

The main output is a combined plot showing the encoded dimension importance on the left and the input pixel importance
(with directional importance highlighted) on the right..
}
