% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity.R
\name{simplify}
\alias{simplify}
\alias{simplify_outcomes}
\alias{simplify_exposures}
\alias{simplify_mediation}
\title{Simplify a complex formula}
\usage{
simplify(x, ...)

simplify_outcomes(x)

simplify_exposures(x)

simplify_mediation(x)
}
\arguments{
\item{x}{A \code{tm}, \code{fmls}, or \code{formula} object}

\item{...}{For future extensions}
}
\value{
A vector of \code{fmls} objects
}
\description{
Simplify a complex formula
}
\details{
Takes either a \code{tm}, \code{fmls}, object or a \code{formula} object as its main
argument. If a \code{fmls} object is given, then the attributes/properties of
those objects are maintained.
}
