% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific-aims.R
\name{specific_aims}
\alias{specific_aims}
\title{Specific Aims}
\usage{
specific_aims(...)
}
\value{
An \code{aims} object that is a collection of the related individual \code{aim} objects from a project.
}
\description{
Every research project, particularly proposals, starts with a specific aims page, which contains the aims and sub-aims intended to be studied along with the corresponding hypothesis. The \code{specific_aims()} function is a method to collect all the related \code{aim} objects from a project to help track, update, and manage the research project.
}
