% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studies.R
\name{studies}
\alias{studies}
\title{Initialize a Framework for studiesing Hypotheses}
\usage{
studies(...)
}
\description{
Initialize a Framework for studiesing Hypotheses
}
\seealso{
Other studies: 
\code{\link{add_hypothesis}()},
\code{\link{build_studies}()}
}
\concept{studies}
