% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{study}
\alias{study}
\title{Initialize a Framework for Studying Hypotheses}
\usage{
study(...)
}
\description{
Initialize a Framework for Studying Hypotheses
}
\seealso{
Other studies: 
\code{\link{add_hypothesis}()},
\code{\link{build_study}()}
}
\concept{studies}
