% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{summary.model_map}
\alias{summary.model_map}
\title{model_map Summary}
\usage{
\method{summary}{model_map}(object, ...)
}
\arguments{
\item{object}{a \code{model_map} object}

\item{...}{further arguments passed to or from other methods}
}
\description{
model_map Summary
}
