% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{summary.model_table}
\alias{summary.model_table}
\title{model_table Summary}
\usage{
\method{summary}{model_table}(object, ...)
}
\arguments{
\item{object}{a \code{model_table} object}

\item{...}{further arguments passed to or from other methods}
}
\description{
model_table Summary
}
