% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{summary.study}
\alias{summary.study}
\title{Study Summary}
\usage{
\method{summary}{study}(object, ...)
}
\arguments{
\item{object}{a \code{study} object}

\item{...}{further arguments passed to or from other methods}
}
\description{
Study Summary
}
