% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt-linear.R
\name{tbl_linear}
\alias{tbl_linear}
\title{Table of linear models}
\usage{
tbl_linear(
  object,
  data,
  outcomes = formula(),
  terms = formula(),
  adjustment = formula(),
  columns = list(beta ~ "Estimate", conf ~ "95\% CI", p ~ "P value"),
  ...
)
}
\arguments{
\item{object}{A \verb{<mdl_tbl>} object with the required models. It must also
contain the original dataset used to create the models of interest. Please
see \code{\link[=attach_data]{attach_data()}} for details.}

\item{data}{The original dataset used to create the models of interest. The
variables in the model must be contained within this dataset. This may also
be attached}

\item{outcomes}{A \verb{<formula>} or list of formulas selecting the outcome
variables of interest. The \strong{LHS} is always the name of the variable that
will be selected. The \strong{RHS} is the potential label for the output table.}

\item{terms}{A \verb{<formula>} or list of formulas selecting the model terms that
should be used. The \strong{LHS} is always the name of the variable that will be
selected. The \strong{RHS} is the potential label for the output table.}

\item{columns}{Additional columns that help to describe the subgroup models.
At least one column should be selected from this list. The sequence listed
will reflect the sequence shown in the table. The current options are:
\itemize{
\item beta = point estimate value, such as odds ratio or hazard ratio
\item conf = inclusion of the confidence interval (presumed to be ~95\%-ile)
\item n = number of observations in each group or subset
\item p = p_value for model or interaction term
}

For example: \verb{list(beta ~ "Hazard", conf ~ "95\% CI" n ~ "No.")"}}

\item{distribution}{A \verb{<character>} describes the distribution of the key
exposure terms that are being displayed. Can either be \emph{continuous},
\emph{dichotomous}, or \emph{categorical}.}
}
\description{
A \verb{<mdl_tbl>} object of linear or generalized linear models. This serves as a
workhorse for a majority of simple regression models and allows them to be
displayed using the \code{{gt}} package format.
}
