% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{tbls}
\alias{tbls}
\title{Extensions of \code{{gt}} package for multiple models}
\arguments{
\item{object}{A \verb{<mdl_tbl>} object with the required models. It must also
contain the original dataset used to create the models of interest. Please
see \code{\link[=attach_data]{attach_data()}} for details.}

\item{data}{The original dataset used to create the models of interest. The
variables in the model must be contained within this dataset. This may also
be attached}

\item{outcomes}{A \verb{<formula>} or list of formulas selecting the outcome
variables of interest. The \strong{LHS} is always the name of the variable that
will be selected. The \strong{RHS} is the potential label for the output table.}

\item{exposures}{A \verb{<formula>} or list of formulas of strata that should be
evaluated, with the \strong{LHS} referring to the strata and the \strong{RHS}
referring to its label.}

\item{interactions}{A \verb{<formula>} or list of formulas of the interaction
terms that should be evaluated, with the \strong{LHS} referring to the term and
the \strong{RHS} referring to its label. Currently only supports binary
interaction terms.}

\item{strata}{A \verb{<formula>} or list of formulas of strata that should be
evaluated, with the \strong{LHS} referring to the strata and the \strong{RHS}
referring to its label.}

\item{level_labels}{A \verb{<formula>} or list of formulas where each list-element
is a formula with the \strong{LHS} reflecting either the variable to re-label or
a specific level, and the \strong{RHS} reflecting what the new level should be
called (for display). If there are conflicting labels, the most recent will
be used.

For example, \verb{list(am ~ c("Manual", "Automatic")} would take, from the
\code{mtcars} dataset, the \code{am} variable, which consists of \code{c(0, 1)}, and
relabel them in the order described. They are sorted in ascendinng order
prior to re-labeling.

The alternative approach is to use the specific level itself and have it
re-labeled. \code{list(0 ~ "Absent")} would take all levels that are zero, and
change their value.}

\item{terms}{A \verb{<formula>} or list of formulas selecting the model terms that
should be used. The \strong{LHS} is always the name of the variable that will be
selected. The \strong{RHS} is the potential label for the output table.}

\item{columns}{Additional columns that help to describe the subgroup models.
At least one column should be selected from this list. The sequence listed
will reflect the sequence shown in the table. The current options are:
\itemize{
\item beta = point estimate value, such as odds ratio or hazard ratio
\item conf = inclusion of the confidence interval (presumed to be ~95\%-ile)
\item n = number of observations in each group or subset
\item p = p_value for model or interaction term
}

For example: \verb{list(beta ~ "Hazard", conf ~ "95\% CI" n ~ "No.")"}}
}
\description{
The related functions in this package combine multiple models to help show
the relationship between variables.
}
\details{
The extensions that combine the modeling functions of the \code{{vlndr}} package
and table-making functions of the \code{{gt}} package require some conventions to
simplify the process. These conventions are listed and explained below.
}
\section{Formula inputs}{
When providing pluralized arguments, such as for \strong{outcomes}, \strong{exposures},
\strong{interactions}, and \strong{strata}, both the variable itself and its potential
label are of interest. These \emph{labeled formulas} are handled generally by the
function \code{\link[=labeled_formulas_to_named_list]{labeled_formulas_to_named_list()}} The input types are (in order of
preference/informativity):
\itemize{
\item \verb{<list>} of \verb{<formula>} objects: The \strong{LHS} is always the name of the variable that will be selected. The \strong{RHS} is the potential label for the output table.
\item \verb{<formula>} object
\item \verb{<character>} vector
}
}

