% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms.R
\name{term_components}
\alias{term_components}
\alias{roles}
\title{Retrieve components of \code{tm} objects}
\usage{
roles(x)
}
\arguments{
\item{x}{A vector \code{tm} objects}
}
\description{
Family of related functions that extract components of \code{tm} objects.
}
