% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidy_tests}
\alias{tidy_tests}
\title{Tidy a list of \code{htest} or model objects}
\usage{
tidy_tests(.fits, conf.int = TRUE, conf.level = 0.95, exponentiate = TRUE)
}
\arguments{
\item{.fits}{List of model or \code{htest} objects that have been fitted}

\item{conf.int}{Logical indicating whether or not to include a confidence
interval in the tidied output. Defaults to \code{FALSE}.}

\item{conf.level}{The confidence level to use for the confidence interval
if \code{conf.int = TRUE}. Must be strictly greater than 0 and less than 1.
Defaults to 0.95, which corresponds to a 95 percent confidence interval.}
}
\value{
Returns a list of tidy objects in the form of tibbles of parameters
}
\description{
Accepts a list of objects that can be tidied via
[\code{\link[broom:reexports]{broom::tidy()}}]
}
