% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framework.R
\name{tie_hypothesis_to_data}
\alias{tie_hypothesis_to_data}
\title{Tie Hypothesis to Data}
\usage{
tie_hypothesis_to_data(framework, hypothesis, name)
}
\description{
Sets the status of the \code{framework} object to tie the hypothesis to the data it will be tested against.
}
\concept{internals}
