% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis-helpers.R
\name{update_hypothesis}
\alias{update_hypothesis}
\title{Update a Hypothesis}
\usage{
update_hypothesis(hypothesis, ...)
}
\arguments{
\item{hypothesis}{A \code{hypothesis} object}
}
\value{
A \code{hypothesis} object
}
\description{
A \code{hypothesis} object can be modified to generate an updated hypothesis,
which can then be placed into a \code{framework} as needed. The \strong{strata} and
\strong{test_opts} arguments can only be updated if the \strong{data} and \strong{test}
options are also given, respectively.
}
