% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimp_encoded.R
\name{vimp_encoded}
\alias{vimp_encoded}
\title{Calculate Variable Importance for Encoded Dimensions}
\usage{
vimp_encoded(
  encoder,
  test_data,
  test_labels,
  autoencoder,
  classes,
  num_permutations = 4,
  errorMetric = "MSE"
)
}
\arguments{
\item{encoder}{A Keras encoder model.}

\item{test_data}{A matrix or array of test data.}

\item{test_labels}{A vector of test labels.}

\item{autoencoder}{A Keras autoencoder model.}

\item{classes}{A vector of class labels to calculate importance for.}

\item{num_permutations}{Integer. Number of permutations for importance calculation. Default is 4.}

\item{errorMetric}{Character. The error metric to use. Default is "MSE".}
}
\value{
A data frame with columns for Class, EncodedDimension, and Importance.
}
\description{
This function calculates the variable importance for each encoded dimension
of an autoencoder model using permutation importance.
}
