% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimp_input.R
\name{vimp_input}
\alias{vimp_input}
\title{Calculate Variable Importance for Encoder Input}
\usage{
vimp_input(encoder, test_data, num_permutations = 4, metric = "MSE")
}
\arguments{
\item{encoder}{A Keras encoder model.}

\item{test_data}{A matrix or data frame containing the test data.}

\item{num_permutations}{Integer. The number of times to permute each feature. Default is 4.}

\item{metric}{Character. The error metric to use. Default is "MSE" (Mean Squared Error).}
}
\value{
A list containing:
  \item{Vimp}{A list of data frames, each representing the variable importance for one dimension of the encoded output.}
}
\description{
This function calculates the variable importance for each input feature of an encoder
using permutation importance. It permutes each feature multiple times and measures
the impact on the encoder's output.
}
