% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimpDecoded.R
\name{vimpDecodedNew}
\alias{vimpDecodedNew}
\title{Variable Importance of Decoded Images in Autoencoder}
\usage{
vimpDecodedNew(encoderModel, num_permutations = 4, errorMetric = "MSE")
}
\arguments{
\item{encoderModel}{A list containing components of the trained autoencoder model.
This should include the encoder, test data (x_test, y_test), and the autoencoder model itself.}

\item{num_permutations}{An integer indicating the number of permutations to be used
for each dimension in the importance calculation. Default is 4.}

\item{errorMetric}{A character string specifying the metric to be used for calculating error.
Default is 'MSE' (Mean Squared Error).}
}
\value{
If `plot = TRUE`, returns a ggplot object visualizing the importance of each encoded dimension
        for each digit. If `plot = FALSE`, returns a data frame with columns 'Digit', 'EncodedDimension',
        and 'Importance'.
}
\description{
This function calculates the importance of each encoded dimension in an autoencoder model
for each digit in the MNIST dataset by using a permutation approach.
}
\examples{
# Assuming 'ae' is a trained autoencoder model on MNIST data
# vimpDecodedNew(ae, num_permutations = 5, plot = TRUE)

}
