% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlndr-package.R
\docType{package}
\name{vlndr-package}
\alias{vlndr}
\alias{vlndr-package}
\title{vlndr package}
\description{
Organize many models and hypotheses, generate directed acyclic graphs from these models, and cut out variables in a causality-based approach to analyses. The package consists of formulating hypothesis, testing the relationships, and analyzing the parameters. In causality-based studies, the focus is on understanding and contextualizing variables. This package provides methods to understand the relationships between outcomes, exposures, and other covariates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/asshah4/vlndr}
  \item Report bugs at \url{https://github.com/asshah4/vlndr/issues}
}

}
\author{
\strong{Maintainer}: Anish Shah \email{ashah282@uic.edu} (\href{https://orcid.org/0000-0002-9729-1558}{ORCID})

}
\keyword{internal}
