% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adding.R
\name{add_data}
\alias{add_data}
\alias{add_data.hypothesis}
\alias{add_data.framework}
\title{Adding Data}
\usage{
add_data(x, ...)

\method{add_data}{hypothesis}(x, .data, .strata = NULL)

\method{add_data}{framework}(x, name, .data, .strata = NULL)
}
\arguments{
\item{x}{Either a \code{hypothesis} or \code{framework} object}

\item{.data}{Data set to be linked to this hypothesis. The argument can
either be a character string or a data frame, and defaults to NULL. If
\code{.data} is a character string, it assumes that a data frame will be added
later (or is already available in the \code{framework} object). If the \code{.data}
is a data frame, it takes the name of the given data. It is not necessary
to include a data argument, but the hypothesis cannot be tested until a
data set is provided. This allows for flexibility and lightweight
manipulation of the \code{hypothesis} object.}

\item{.strata}{How the data should be split or stratified. References the
name of the data given in that the models will be with fit against,
splitting the data into subsets. This helps to perform hypothesis testing
on subsets or strata of the data. It defaults to NULL (which means the full
data will be used). This argument will only be incorporated if \code{.data} is
present.}

\item{name}{Character vector of which hypothesis the data should be added to.}
}
\value{
Returns a \code{hypothesis} or \code{framework} object with the prescribed data added back on, which can be seen by calling \code{print(framework)} or \code{attributes(hypothesis)$data}.
}
\description{
If a hypothesis does not have a data, it can be added subsequently. This can be done in a pipe, either to the hypothesis itself, or to a framework, to allow for flexibility.
}
\concept{helpers}
