% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_hypothesis.R
\name{add_hypothesis}
\alias{add_hypothesis}
\title{Add a Hypothesis for a Model Map}
\usage{
add_hypothesis(
  model_map,
  hypothesis,
  name = deparse(substitute(hypothesis)),
  ...
)
}
\arguments{
\item{model_map}{Object of \code{model_map} class}

\item{hypothesis}{Object of \code{hypothesis} class}

\item{name}{Name of the \code{hypothesis} object, which defaults to the name of
the \code{hypothesis} object itself}

\item{...}{For extensibility}
}
\value{
A \code{model_map} object that has had a \code{hypothesis} added
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Takes a \code{hypothesis} object and adds it to a \code{model_map}, allowing multiple
potential hypotheses to be put together for eventual analysis and comparison.
This involves the deconstructing of the hypothesis into its individual parts
and recomposition of the theorized formula.
}
\concept{hypotheses maps}
