% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framework.R
\name{add_hypothesis}
\alias{add_hypothesis}
\title{Add a Hypothesis to the Framework}
\usage{
add_hypothesis(
  framework,
  hypothesis,
  name = deparse(substitute(hypothesis)),
  ...
)
}
\description{
Add a Hypothesis to the Framework
}
\seealso{
Other frameworks: 
\code{\link{framework}()}
}
\concept{frameworks}
