% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_hypothesis.R
\name{add_hypothesis}
\alias{add_hypothesis}
\title{Add a Hypothesis to the Project}
\usage{
add_hypothesis(
  project,
  name,
  formula,
  combination = "direct",
  test,
  which_data = NULL,
  .strata = NA,
  .stage = "hypothesis",
  ...
)
}
\arguments{
\item{project}{Object of class \code{project}}

\item{name}{Character string to identify this arm. Must be unique and not
previously specified.}

\item{formula}{Formula showing relationship of outcomes and predictors, and
is essentially the hypothesis. It allows for complex formulas e.g. multiple
predictors, multiple exposures, and multiple covariates, that can be
organized into individual formulas based on additional modifiers. Each RHS
term of the formula is considered an outcome variable, and is analyzed as a
single outcome. Each LHS term is consider a predictor, and can be modified
as below:
\itemize{
\item \code{exp()} is placed around a term to define as an independent exposure,
which will be placed in separate formulas from any other term marked as an
exposure
\item \code{cov()} is placed around a term for any predictors that should be
maintained/fixed in all models, which can include complex terms, such as
mixed effects
}

Of note, these formula modifiers do not apply to \code{htest} objects yet.

For example, the equation below describes two independent exposures "x1"
and "x2" that should be conditionally regressed for every level of "z"

\deqn{y ~ exp(x1) + exp(x2) + x3 + x4 + cov((1 | z))}}

\item{combination}{The building pattern for how to put together the overall
plan. It defines variable relationships that will be used. The options for
the \code{combination} currently include:
\itemize{
\item \code{direct} will define the relationship as y ~ x
\item \code{sequential} will define the relationship as y ~ x1, y ~ x1 + x2
\item \code{parallel} will define the relationship as y ~ x1, y ~ x2
}}

\item{test}{The testing that will be used, defined as how the analysis or
hypothesis testing should be performed. There are several options on which
specification to use, anywhere from regression modeling to inferential
statistics. The test applies to the collection of hypotheses.
\itemize{
\item A model specified by the \href{https://parsnip.tidymodels.org}{\code{parsnip} package}, which includes the mode and #'   computational engine
\item A statistical test, such as a \code{t.test}, which may require additional
parameters. These can be given as additional, unmatched arguments. This
option currently supports only hypothesis tests, of class \code{htest}.
}}

\item{which_data}{Name of "row" or dataset that should be analyzed. Defaults
to the most recent dataset added.}

\item{.strata}{How the data should be split or stratified. References the
name of the data given in \code{set_data()} that the models will be with fit
against, splitting the data into subsets. This helps to perform hypothesis
testing on subsets or strata of the data. It defaults to NULL (which means
the full data will be used) \strong{experimental}}

\item{.stage}{Internal marker of project status.}

\item{...}{This should reflect the additional parameters that may need to be
given to the \code{test} argument, such as \code{paired = TRUE} for \code{t.test()}. The
additional parameters must be named to allow them to be passed
successfully.}
}
\value{
\code{project} object
}
\description{
Adding an "arm" to the project reflects a series of hypothesis generated
around the core data that was set in the preceding function. The hypothesis
formula allows for modifiers to both the formula itself, as well as \emph{how} it
should be applied to the data set.
}
