% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framework.R
\name{add_hypothesis}
\alias{add_hypothesis}
\title{Add a Hypothesis to the Framework}
\usage{
add_hypothesis(
  framework,
  hypothesis,
  name = deparse(substitute(hypothesis)),
  ...
)
}
\arguments{
\item{framework}{Object of class \code{framework}}

\item{hypothesis}{Object of class \code{hypothesis} (which may or may not include
data already added)}

\item{name}{Name of the \code{hypothesis} object, which defaults to the name of
the \code{hypothesis} object itself}
}
\value{
A \code{framework} object that has had a \code{hypothesis} added
}
\description{
Takes a \code{hypothesis} object and adds it to a \code{framework}, allowing multiple
potential hypotheses to be put together for eventual analysis and comparison.
}
\seealso{
Other frameworks: 
\code{\link{build_frames}()},
\code{\link{fit_models}()},
\code{\link{framework}()}
}
\concept{frameworks}
