% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{add_hypothesis}
\alias{add_hypothesis}
\title{Add a Hypothesis to the Study}
\usage{
add_hypothesis(study, hypothesis, name = deparse(substitute(hypothesis)), ...)
}
\arguments{
\item{study}{Object of class \code{study}}

\item{hypothesis}{Object of class \code{hypothesis} (which may or may not include
data already added)}

\item{name}{Name of the \code{hypothesis} object, which defaults to the name of
the \code{hypothesis} object itself}
}
\value{
A \code{study} object that has had a \code{hypothesis} added
}
\description{
Takes a \code{hypothesis} object and adds it to a \code{study}, allowing multiple
potential hypotheses to be put together for eventual analysis and comparison.
}
\seealso{
Other studies: 
\code{\link{build_study}()},
\code{\link{study}()}
}
\concept{studies}
