% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arm.R
\name{add_arm}
\alias{add_arm}
\title{Add Arms to the \code{octomod}}
\usage{
add_arm(
  octomod,
  title = NULL,
  f = NULL,
  exposure = NULL,
  pattern = "direct",
  approach = NULL,
  ...
)
}
\arguments{
\item{octomod}{Object of class \code{octomod}}

\item{title}{Character string to identify this arm}

\item{f}{Formula showing relationship of outcomes and predictors, and is the essentially the hypothesis.}

\item{exposure}{Variable(s) that are forced to be maintained in every model as
a predictor.}

\item{pattern}{The variable relationship approach that will be used. The
options for the \code{pattern} are:
\itemize{
\item \code{direct} will define the relationship as y ~ x
\item \code{sequential} will define the relationship as y ~ x1, y ~ x1 + x2
\item \code{parallel} will define the relationship as y ~ x1, y ~ x2
}}

\item{approach}{This describes the analysis plan that should be used. There
are several options on which specification to use, anywhere from regression
modeling to inferential statistics. The approach applies to the entirety of
the growing arm.
\itemize{
\item A model specified by the \href{https://parsnip.tidymodels.org}{\code{parsnip} package}, which includes the mode and
computational engine
\item A statistical test, such as a \code{t.test}, which may require additional
parameters. These can be given as additional, unmatched arguments. This
option currently supports only hypothesis tests, of class \code{htest}.
}}

\item{...}{This should reflect the additional parameters that may need to be
given to the \code{approach} argument, such as \code{paired = TRUE} for \code{t.test}. The
additional parameters must be named to allow them to be passed
successfully.}
}
\value{
An \code{octomod} object with arms attached
}
\description{
Add a new hypothesis arm to the \code{octomod} object. The hypothesis collected
here are all directed at the core dataset specified by the \code{add_core()}
function.
}
\details{
This step is the most crucial, in that the the hypothesis and relationship
between variables is defined at this stage.
}
\examples{
library(magrittr)

om <-
  octomod() \%>\%
  add_core(mtcars) \%>\%
  add_arm(
    title = "Horsepower",
    f = disp ~ hp,
    pattern = "direct",
    approach = "t.test",
    paired = TRUE
  )

}
