% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullet.R
\name{bullet}
\alias{bullet}
\title{Create bullets to be loaded}
\usage{
bullet(f, exposure = NULL, approach = "sequential", model, ...)
}
\arguments{
\item{f}{Formula showing relationship of outcomes to predictors}

\item{exposure}{Variable(s) that is forced to be maintained in every model as
a predictor.}

\item{approach}{The modeling approach that will be used. The options are:
\itemize{
\item \code{sequential} will build y ~ x1, y ~ x1 + x2 models
\item \code{parallel} will build y ~ x1, y ~ x2 models
}}

\item{model}{A model specified by the \href{https://parsnip.tidymodels.org/articles/parsnip_Intro.html}{\code{parsnip} package}, which
includes the mode and computational engine}

\item{...}{For extensibility}
}
\value{
Object of type \code{bullet} that contains hypothesis characteristics
needed for analysis
}
\description{
Create bullets to be loaded
}
\examples{
library(parsnip)
f <- mpg + cyl ~ wt + hp + gear
lm_mod <- linear_reg() \%>\% set_engine("lm")
b <- bullet(f, exposure = "wt", approach = "sequential", model = lm_mod)

}
