% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullet.R
\name{bullet}
\alias{bullet}
\title{Create bullets to be loaded}
\usage{
bullet(f, exposure = NULL, approach = "sequential", model, ...)
}
\arguments{
\item{f}{Formula showing relationship of outcomes and predictors}

\item{exposure}{Variable(s) that is forced to be maintained in every model as
a predictor.}

\item{approach}{The modeling approach that will be used. The options are:
\itemize{
\item \code{sequential} will build y ~ x1, y ~ x1 + x2 models
\item \code{parallel} will build y ~ x1, y ~ x2 models
}}

\item{model}{A model specified by the \href{https://parsnip.tidymodels.org/articles/parsnip_Intro.html}{\code{parsnip} package}, which
includes the mode and computational engine}

\item{...}{For extensibility}
}
\value{
Object of type \code{bullet} that contains the hypothesis and the
characteristics of how to put the analysis together. needed for analysis
}
\description{
A \code{bullet} object is a pre-specified hypothesis, with a formula
that helps to identify the individual variables and their roles. It
requires the use of a model specification from \code{parsnip} (or any other
similarly specified models). It also determines what sequence of models to
put together.
}
\examples{
library(parsnip)
f <- mpg + cyl ~ wt + hp + gear
lm_mod <- linear_reg() \%>\% set_engine("lm")
b <- bullet(f, exposure = "wt", approach = "sequential", model = lm_mod)

}
