% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_findings.R
\name{collect_findings}
\alias{collect_findings}
\title{Collect the Important Findings from a Project}
\usage{
collect_findings(
  project,
  which_tests = NULL,
  tidy = TRUE,
  .stage = "results",
  ...
)
}
\arguments{
\item{project}{Object of class \code{project}}

\item{which_tests}{Character vector of hypothesis names to be collected.
Defaults to all tests that were run.}

\item{tidy}{Returns findings after being processed by \code{\link[broom:reexports]{broom::tidy()}},
defaulting to TRUE. If the original model fit is required, change to FALSE.}

\item{.stage}{Internal marker of workflow progress.}

\item{...}{For extensibility}
}
\value{
Returns a \code{tibble} that includes the name of the hypothesis, the test
number, the outcome variable, the terms, and model or test parameters as
defined by the \code{\link[broom:reexports]{broom::tidy()}} function. If requested, can instead return
the original model fit instead of tidy parameters.
}
\description{
This function simplifies the collection of the model fit and test results
from a function. This should be called after the project has been fitted with
\code{build_models()}.
}
