% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_findings.R
\name{collect_findings}
\alias{collect_findings}
\title{Collect the Important Findings from a Project}
\usage{
collect_findings(project, ...)
}
\arguments{
\item{project}{Object of class \code{project}}

\item{...}{For extensibility.}
}
\value{
\code{project} object
}
\description{
This function simplifies the collection of the model fit and test results from a function. This should be called after the project has been fitted with \code{build_models()}.
}
