% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{add_core}
\alias{add_core}
\alias{update_core}
\alias{remove_core}
\title{Add Core Data to the \code{octomod}}
\usage{
add_core(octomod, core)

update_core(octomod, new_core, ...)

remove_core(octomod, ...)
}
\arguments{
\item{octomod}{Object of class \code{octomod}}

\item{core}{Data frame or tibble object}

\item{new_core}{A modified or new data frame to be used}

\item{...}{For extensibility}
}
\value{
\code{octomod} object with core data attached
}
\description{
Add the shared dataset for this research question to the central \code{octomod}
structure. Whenever this is used, it will replace the prior or old data.

\code{add_core()} adds a new dataset to the \code{octomod} object. It checks to see if
data is already present.

\code{update_core()} changes the old dataset from the \code{octomod} object to an
updated dataset. It does not re-do the hypothesis testing.

\code{remove_core()} removes the data from the \code{octomod} object. This is usually
not needed.
}
\examples{
library(magrittr)

octomod() \%>\%
  add_core(mtcars)

}
