% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forging-tools.R
\name{draw}
\alias{draw}
\title{Draw out components from the forge}
\usage{
draw(x, f)
}
\arguments{
\item{x}{A model \code{forge} object}

\item{f}{A formula or formula_archetype object}
}
\description{
Draw out components from the forge
}
