% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw_hypothesis}
\alias{draw_hypothesis}
\title{Draw Hypothesis onto Map}
\usage{
draw_hypothesis(study, hypothesis, name = deparse(substitute(hypothesis)), ...)
}
\arguments{
\item{study}{Object of class \code{study}}

\item{hypothesis}{Object of class \code{hypothesis} (which may or may not include
data already added)}

\item{name}{Name of the \code{hypothesis} object, which defaults to the name of
the \code{hypothesis} object itself}

\item{...}{For extensibility}
}
\value{
A \code{study} object that has had a \code{hypothesis} added
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Takes a \code{hypothesis} object and adds it to a \code{study}, allowing multiple
potential hypotheses to be put together for eventual analysis and comparison.
This involves the deconstructing of the hypothesis into its individual parts
and recomposition of the theorized formula.

Simultaneously, takes the \code{hypothesis} objects that were added to the \code{study}
and decomposes them into specific \strong{paths} that are used to help define
relationships between variables. These paths are stored in the \code{study} itself
in the form of a \code{data.frame}. This is currently experimental in that the
directionality, relationships, and patterns are intended to be used to help
identify certain variables for future modeling, but the implementation is not
yet complete.
}
\seealso{
Other studies: 
\code{\link{construct_map}()},
\code{\link{study}()}
}
\concept{studies}
