% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{estimate_interaction}
\alias{estimate_interaction}
\title{Estimating interaction effect estimates}
\usage{
estimate_interaction(object, exposure, interaction, conf_level = 0.95, ...)
}
\value{
A \verb{<data.frame>} with \code{n = 2} rows (for the presence or absence of
the interaction term).
}
\description{
When using categorical interaction terms in a \verb{<mdl_tbl>} object, estimates
on interaction terms and their confidence intervals can be evaluated. The
effect of interaction on the estimates is based on the levels of interaction
term. The estimates and intervals can be derived through the
\code{estimate_interaction()} function. The approach is based on the method
described by Figueiras et al. (1998).
}
\details{
The \code{estimate_interaction()} requires a \verb{<mdl_tbl>} object that is a
\emph{single row in length}. Filtering the \verb{<mdl_tbl>} should occur prior to passing
it to this function. Additionally, this function assumes the interaction term
is binary. If it is categorical, the current recommendation is to use dummy
variables for the corresponding levels prior to modeling.
}
\references{
A. Figueiras, J. M. Domenech-Massons, and Carmen Cadarso, 'Regression models:
calculating the confidence intervals of effects in the presence of
interactions', Statistics in Medicine, 17, 2099-2105 (1998)
}
