% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{expand_formula}
\alias{expand_formula}
\title{Expand Formulae for a Hypothesis}
\usage{
expand_formula(formula, combination, ...)
}
\arguments{
\item{formula}{Formula showing relationship of outcomes and predictors, and
is essentially the hypothesis. It allows for complex formulae e.g. multiple
predictors, multiple exposures, and multiple covariates, that can be
organized into individual formulae based on additional modifiers. Please
see the  for further documentation.}

\item{combination}{The building pattern for how to put together the overall
plan. It defines variable relationships that will be used. The options for
the \code{combination} currently include:
\itemize{
\item \code{direct} will define the relationship as y ~ x
\item \code{sequential} will define the relationship as y ~ x1, y ~ x1 + x2
\item \code{parallel} will define the relationship as y ~ x1, y ~ x2
}}

\item{...}{Not currently used}
}
\value{
Table of formulae and extracted terms
}
\description{
This is an internal function that helps to create the appropriate formula
combinations for adding hypotheses.
}
\section{Formulae}{


When supplying a \code{formula} object, each RHS term of the formula is
considered an outcome variable, and is analyzed as a single outcome. Each
LHS term is consider a predictor, and can be modified as below:
\itemize{
\item \code{X()} is placed around a term to define as an independent exposure, which
will be placed in separate formulae from any other term marked as an
exposure
\item \code{F()} is placed around a term for any predictors that should be
maintained/fixed in all models, which can include complex terms, such as
mixed effects
}

For example, the equation below describes two independent exposures "x1"
and "x2" that should be conditionally regressed for every level of "z"

\deqn{y ~ X(x1) + X(x2) + x3 + x4 + F((1 | z))}
}

