% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressions.R, R/hypothesize.R
\name{expand_formula}
\alias{expand_formula}
\title{Expand Formulae for a Hypothesis}
\usage{
expand_formula(formula, combination, table = FALSE, ...)

expand_formula(formula, combination, table = FALSE, ...)
}
\arguments{
\item{formula}{Formula showing relationship of outcomes and predictors, and
is essentially the hypothesis. It allows for complex formulae e.g. multiple
predictors, multiple exposures, and multiple covariates, that can be
organized into individual formulae based on additional modifiers. Please
see the  for further documentation.}

\item{combination}{The building pattern for how to put together the overall
plan. It defines variable relationships that will be used. The options of
\code{combination} will lead to the following patterns:
\itemize{
\item \code{direct} will not modify the original equation pattern
}

\deqn{y ~ x1 + x2}
\itemize{
\item \code{sequential} will create a list of sequentially adjusted formula
}

\deqn{y ~ x1} \deqn{y ~ x1 + x2}
\itemize{
\item \code{parallel} will create a list of formula with parallel predictors
}

\deqn{y ~ x1} \deqn{y ~ x2}}

\item{table}{Determines if the returned value should be a list of formulas or
a table. Defaults to FALSE, leading to a list generation.}

\item{...}{Not currently used}
}
\value{
List or table of formulae and extracted terms (to be determined by
the \strong{table} argument).

List or table of formulae and extracted terms (to be determined by
the \strong{table} argument).
}
\description{
This is an internal function that helps to create the appropriate formula
combinations for adding hypotheses.

This is an internal function that helps to create the appropriate formula
combinations for adding hypotheses.
}
\section{Formulae}{


When supplying a \code{formula} object, each RHS term of the formula is
considered an outcome variable, and is analyzed as a single outcome. Each
LHS term is consider a predictor, and can be modified as below:
\itemize{
\item \code{X()} is placed around a term to define as an independent exposure, which
will be placed in separate formulae from any other term marked as an
exposure
\item \code{F()} is placed around a term for any predictors that should be
maintained/fixed in all models, which can include complex terms, such as
mixed effects
\item \code{C()} is placed around a term for any predictors that are a potential
confounder, which does not yet change the relationship of how formulas are
built, but is stored for future methods to be used with \code{\link[=hypothesize]{hypothesize()}}
}

For example, the equation below describes two independent exposures "x1"
and "x2" that should be conditionally tested for every level of "z"

\deqn{y ~ X(x1) + X(x2) + x3 + x4 + F((1 | z))}



When supplying a \code{formula} object, each RHS term of the formula is
considered an outcome variable, and is analyzed as a single outcome. Each
LHS term is consider a predictor, and can be modified as below:
\itemize{
\item \code{X()} is placed around a term to define as an independent exposure, which
will be placed in separate formulae from any other term marked as an
exposure
\item \code{F()} is placed around a term for any predictors that should be
maintained/fixed in all models, which can include complex terms, such as
mixed effects
\item \code{C()} is placed around a term for any predictors that are a potential
confounder, which does not yet change the relationship of how formulas are
built, but is stored for future methods to be used with \code{\link[=hypothesize]{hypothesize()}}
}

For example, the equation below describes two independent exposures "x1"
and "x2" that should be conditionally tested for every level of "z"

\deqn{y ~ X(x1) + X(x2) + x3 + x4 + F((1 | z))}
}

