% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_models}
\alias{extract_models}
\title{Extract models from a study}
\usage{
extract_models(study, which_ones = NULL, tidy = TRUE, ...)
}
\arguments{
\item{study}{A \code{study} object that contains \code{hypothesis} objects that have
been run}

\item{which_ones}{Vector of names of hypotheses that should be extracted and
distilled. Will only be applied to models that have been fit. Defaults to
NULL, for which all tests will be returned.}

\item{tidy}{Logical value for if the returned models should be collected into
a tidy table or as a list of raw model/test fits. Defaults to TRUE.}

\item{...}{For extensibility}
}
\value{
A table that has either raw or tidy lists of model fits
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function serves as a non-pipeable extraction function to obtain
aggregate components from a \code{study} object. Its designed to have a simple
call interface to allow for retrieval of model objects to be formatted and
displayed.
}
