% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confound.R
\name{find_confounders}
\alias{find_confounders}
\title{Derive Formula With Confounders}
\usage{
find_confounders(model_map, name, delta = 0.1, return_map = FALSE, ...)
}
\arguments{
\item{model_map}{A \code{model_map} object that has been constructed}

\item{name}{Name of a an existing \code{hypothesis} object}

\item{delta}{Percent change to be used as cut-off for considering an
important change in the outcome ~ exposure relationship. Defaults to
\strong{0.10} as supported by the epidemiology literature.}

\item{return_map}{Logical value, defaults to FALSE, on if the original
\code{model_map} should be returned when this function is called. This is a
developer feature, mainly for internal use (as is the function itself). It
returns the modified hypotheses as a list in position \verb{[[1]]}, and the
updated model_map in position \verb{[[2]]}.}

\item{...}{For extensibility}
}
\value{
A list of \code{hypothesis} objects that have been modified (and can
optionally return original, updated \code{model_map} as well)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Removes confounders from a hypothesis object based on different approaches.
\itemize{
\item \strong{sequential} = The model is analyzed for each additional covariate
added, looking for changes in effect size defined by the variable \code{delta}.
If the change is significant, the covariate is maintained as a potential
confounder, otherwise removed.
\item \strong{parallel} = The parallel models are analyzed for the effect of each
individual covariate that was tested against the \emph{outcome ~ exposure}
relationship. The change in effect size of that relationship is examined
for being of size \code{delta}. If so, the covariate is retained as a potential
confounder (or otherwise removed).
}
}
\concept{confounders finders}
