% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesize.R
\name{hypothesize}
\alias{hypothesize}
\alias{hypothesize.formula}
\alias{hypothesize.default}
\title{Draw A Hypothesis}
\usage{
hypothesize(h, ...)

\method{hypothesize}{formula}(h, combination, test, test_opts = NULL, data, strata = NULL)

\method{hypothesize}{default}(h, ...)
}
\description{
Draw A Hypothesis
}
