% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesize.R
\name{hypothesize}
\alias{hypothesize}
\alias{hypothesize.formula}
\title{Create a Hypothesis}
\usage{
hypothesize(x, ...)

\method{hypothesize}{formula}(
  formula,
  combination,
  test,
  test_opts = NULL,
  data = NULL,
  data_name = NULL,
  strata = NULL,
  ...
)
}
\arguments{
\item{x}{General object that can be cast into a hypothesis object. Currently
this accepts the following object types:
\itemize{
\item \code{formula} objects, which will be given additional modifiers as described.
Please see the \code{formula} parameter for further details.
}}

\item{...}{For additional variables based on the generic method invoked.}

\item{formula}{Formula showing relationship of outcomes and predictors, and
is essentially the hypothesis. It allows for complex formulas e.g. multiple
predictors, multiple exposures, and multiple covariates, that can be
organized into individual formulas based on additional modifiers. Each RHS
term of the formula is considered an outcome variable, and is analyzed as a
single outcome. Each LHS term is consider a predictor, and can be modified
as below:
\itemize{
\item \code{X()} is placed around a term to define as an independent exposure,
which will be placed in separate formulas from any other term marked as an
exposure
\item \code{F()} is placed around a term for any predictors that should be
maintained/fixed in all models, which can include complex terms, such as
mixed effects
}

Of note, these formula modifiers do not apply to \code{htest} objects yet.

For example, the equation below describes two independent exposures "x1"
and "x2" that should be conditionally regressed for every level of "z"

\deqn{y ~ X(x1) + X(x2) + x3 + x4 + F((1 | z))}}

\item{combination}{The building pattern for how to put together the overall
plan. It defines variable relationships that will be used. The options for
the \code{combination} currently include:
\itemize{
\item \code{direct} will define the relationship as y ~ x
\item \code{sequential} will define the relationship as y ~ x1, y ~ x1 + x2
\item \code{parallel} will define the relationship as y ~ x1, y ~ x2
}}

\item{test}{The testing that will be used, defined as how the analysis or
hypothesis testing should be performed. There are several options on which
specification to use, anywhere from regression modeling to inferential
statistics. The test applies to the collection of hypotheses.
\itemize{
\item A model specified by the \href{https://parsnip.tidymodels.org}{\code{parsnip} package}, which includes the mode and
computational engine
\item A statistical test, such as a \code{t.test}, which may require additional
parameters. These can be given as additional, unmatched arguments. This
option currently supports only hypothesis tests, of class \code{htest}.
}}

\item{test_opts}{For optional or additional parameters to be given to the
\emph{test} argument. Should be in format of a list of \strong{name = argument}
pairs.}

\item{data}{Data set to be attached to this hypothesis. Not necessary to
create a \code{hypothesis}, and can be specified or added at a later time. The
name of the data set will be used as the reference variable.}

\item{data_name}{Name of the dataset that this hypothesis should be tested
against. The data itself will be used to generate the name if it is given
from the \emph{data} argument. If only a character vector is given, the function
will match the \code{hypothesis} to data already available (or warn if not
available). This flexibility allows for lightweight manipulation of the
\code{hypothesis} object, and the addition of multiple data sets.}

\item{strata}{How the data should be split or stratified. References the
name of the data given in that the models will be with fit
against, splitting the data into subsets. This helps to perform hypothesis
testing on subsets or strata of the data. It defaults to NULL (which means
the full data will be used) \strong{experimental}}
}
\value{
An object of class \code{hypothesis}
}
\description{
\code{hypothesize()} creates a \code{hypothesis} object that can be placed into a
modeling framework. This framework allows for multiple hypotheses to be
modeled, and allows for assessment of confounding variables.
}
