% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesize.R
\name{hypothesize}
\alias{hypothesize}
\alias{hypothesize.formula}
\alias{hypothesize.list}
\alias{hypothesize.default}
\title{Create a Hypothesis}
\usage{
hypothesize(h, ...)

\method{hypothesize}{formula}(h, combination, test, test_opts = NULL, data, strata = NULL, ...)

\method{hypothesize}{list}(h, combination, test, test_opts = NULL, data, strata = NULL, ...)

\method{hypothesize}{default}(h, ...)
}
\arguments{
\item{h}{General variable that can be cast into a hypothesis object. Currently
this accepts the following object types:
\itemize{
\item \code{formula} objects, which  be given additional modifiers as described.
Please see \code{\link[=expand_formula]{expand_formula()}} or the formula section below for details.
\item \code{list} object, which describes the outcomes, exposures, covariates, fixed
variables, and potential confounders in a named list, which allows for
grouped covariates when modeling. Please see the details section for
further explanation.
}}

\item{...}{For additional variables based on the generic method invoked.}

\item{combination}{The building pattern for how to put together the overall
plan. It defines variable relationships that will be used. The options for
the \code{combination} currently include:
\itemize{
\item \code{direct} will define the relationship as y ~ x
\item \code{sequential} will define the relationship as y ~ x1, y ~ x1 + x2
\item \code{parallel} will define the relationship as y ~ x1, y ~ x2
}}

\item{test}{The testing that will be used, defined as how the analysis or
hypothesis testing should be performed. There are several options on which
specification to use, anywhere from regression modeling to inferential
statistics. The test applies to the collection of hypotheses.
\itemize{
\item A model specified by the \href{https://parsnip.tidymodels.org}{\code{parsnip} package}, which includes the mode and
computational engine
\item A statistical test, such as a \code{t.test}, which may require additional
parameters. These can be given as additional, unmatched arguments. This
option currently supports only hypothesis tests, of class \code{htest}.
}}

\item{test_opts}{For optional or additional parameters to be given to the
\emph{test} argument. Should be in format of a list with name-pairs, such as
\code{list(name = argument)} pairs.}

\item{data}{Data set to be linked to this hypothesis. It assumes data is one
of type \code{c("tbl_df", "tbl", "data.frame")}. The name of the data set is
taken from the object itself.}

\item{strata}{How the data should be split or stratified. References the
name of the data given in that the models will be with fit against,
splitting the data into subsets. This helps to perform hypothesis testing
on subsets or strata of the data. It defaults to NULL (which means the full
data will be used). This argument will only be incorporated if \code{data} is
present.}
}
\value{
An object of class \code{hypothesis}
}
\description{
\code{hypothesize()} creates a \code{hypothesis} object that can be placed into a
modeling framework. This framework allows for multiple hypotheses to be
modeled, and allows for assessment of confounding variables.
}
\section{Warning}{


The formula object does not necessarily work for \code{htest} objects as it is
still \strong{experimental}.
}

\section{Formulae}{


When supplying a \code{formula} object, each RHS term of the formula is
considered an outcome variable, and is analyzed as a single outcome. Each
LHS term is consider a predictor, and can be modified as below:
\itemize{
\item \code{X()} is placed around a term to define as an independent exposure, which
will be placed in separate formulae from any other term marked as an
exposure
\item \code{F()} is placed around a term for any predictors that should be
maintained/fixed in all models, which can include complex terms, such as
mixed effects
\item \code{C()} is placed around a term for any predictors that are a potential
confounder, which does not yet change the relationship of how formulas are
built, but is stored for future methods to be used with \code{\link[=hypothesize]{hypothesize()}}
}

For example, the equation below describes two independent exposures "x1"
and "x2" that should be conditionally tested for every level of "z"

\deqn{y ~ X(x1) + X(x2) + x3 + x4 + F((1 | z))}
}

