% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{interaction_estimates}
\alias{interaction_estimates}
\title{Models with interaction variables}
\usage{
interaction_estimates(object, exposure, interaction, conf.level = 0.95, ...)
}
\description{
Models with interaction variables
}
\references{
A. Figueiras, J. M. Domenech-Massons, and Carmen Cadarso, 'Regression models:
calculating the confidence intervals of effects in the presence of
interactions', Statistics in Medicine, 17, 2099-2105 (1998)
}
\keyword{internal}
