% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-table.R
\name{mdl_tbl}
\alias{mdl_tbl}
\alias{model_table}
\alias{is_model_table}
\alias{new_model_table}
\title{Model Tables}
\usage{
mdl_tbl(...)

model_table(...)

is_model_table(x)

new_model_table(
  x = list(),
  formulaMatrix = data.frame(),
  termTable = data.frame(),
  ...
)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function introduces a super class that combines both the \code{list} class
(and its derivative \code{list_of}) and regression models and/or hypothesis tests.
Models that are similar and share certain properties can be combined together
into a \code{mdl_tbl}.
}
