% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-table.R
\name{model_table_helpers}
\alias{model_table_helpers}
\alias{attach_data}
\alias{flatten_models}
\title{Model table helper functions}
\usage{
attach_data(x, data, ...)

flatten_models(x, ...)
}
\arguments{
\item{x}{a \verb{<mdl_tbl>} object}

\item{data}{a \verb{<data.frame>} object that has been used by models}

\item{...}{Arguments to be passed to or from other methods}
}
\description{
These functions are used to help manage the \verb{<mdl_tbl>} object. They allow
for specific manipulation of the internal components, and are intended to
generally extend the functionality of the object.
\itemize{
\item \code{attach_data()}: Attaches a dataset to a \verb{<mdl_tbl>} object
\item \code{flatten_model()}: Flattens a \verb{<mdl_tbl>} object down to its specific parameters
}
}
\section{Attaching Data}{
When models are built, oftentimes the included matrix of data is available
within the raw model, however when handling many models, this can be
expensive in terms of memory and space. By attaching datasets independently
that persist regardless of the underlying models, and by knowing which models
used which datasets, it can be ease to back-transform information.
}

\section{Flattening Models}{
A \verb{<mdl_tbl>} object can be flattened to its specific parameters, their
estimates, and model-level summary statistics.
}

