% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{models}
\alias{models}
\alias{mdl}
\alias{model}
\title{Model prototypes}
\usage{
mdl(
  x = unspecified(),
  formulas = fmls(),
  data_name = NA_character_,
  strata_variable = NA_character_,
  strata_level = NA,
  ...
)

model(
  x = unspecified(),
  formulas = fmls(),
  data_name = NA_character_,
  strata_variable = NA_character_,
  strata_level = NA,
  ...
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
