% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study-helpers.R
\name{modify_study}
\alias{modify_study}
\alias{add_study_formula}
\alias{add_study_path}
\alias{add_study_test}
\alias{add_study_data}
\alias{update_study_strata}
\alias{add_study_status}
\alias{update_study_status}
\title{Study Modifications}
\usage{
add_study_formula(study, hypothesis, name)

add_study_path(study, name)

add_study_test(study, hypothesis, name)

add_study_data(study, hypothesis, name)

update_study_strata(study, hypothesis, name)

add_study_status(
  study,
  hypothesis,
  name,
  run = FALSE,
  error = FALSE,
  path = FALSE,
  origin = NA
)

update_study_status(study, name, ...)
}
\arguments{
\item{study}{Object of \code{study} class}

\item{hypothesis}{Object of \code{hypothesis} class}

\item{name}{Name of \code{hypothesis}, given from parent function}

\item{...}{additional arguments to pass along}
}
\value{
A \code{study} object
}
\description{
Either add, update, or remove components or attributes of a \code{study}
}
\keyword{internal}
