% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{mx}
\alias{mx}
\alias{model_card}
\alias{model_card.lm}
\alias{model_card.glm}
\alias{model_card.model_fit}
\alias{model_card.default}
\title{Decorated Models}
\usage{
model_card(x = unspecified(), ...)

\method{model_card}{lm}(
  x,
  model_name = deparse(substitute(x)),
  term_labels = list(),
  term_roles = list(),
  model_label = character(),
  model_description = character(),
  ...
)

\method{model_card}{glm}(
  x,
  model_name = deparse(substitute(x)),
  term_labels = list(),
  term_roles = list(),
  model_label = character(),
  model_description = character(),
  ...
)

\method{model_card}{model_fit}(
  x,
  model_name = deparse(substitute(x)),
  term_labels = list(),
  term_roles = list(),
  model_label = character(),
  model_description = character(),
  ...
)

\method{model_card}{default}(x = unspecified(), ...)

mx(x = unspecified(), ...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
