% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesize.R
\name{print.hypothesis}
\alias{print.hypothesis}
\title{Print a Hypothesis}
\usage{
\method{print}{hypothesis}(x, ...)
}
\arguments{
\item{x}{A \code{hypothesis} object}

\item{...}{For further arguments to be passed on}
}
\description{
Print a Hypothesis
}
