% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confound.R
\name{reconstruct}
\alias{reconstruct}
\title{Reconstruct a Hypothesis}
\usage{
reconstruct(
  study,
  name,
  new_name = paste0(name, "_cut"),
  approach = "confounding",
  ...
)
}
\arguments{
\item{study}{A \code{study} object that has been constructed}

\item{name}{Name of a an existing \code{hypothesis} object}

\item{new_name}{Name of the new \code{hypothesis} that is created after
modification. Defaults to modifying the original \strong{name} by appending
\verb{*_cut}.}

\item{approach}{String that informs \emph{how} to reconstruct the
hypothesis. Options include:
\itemize{
\item \strong{confounding} = Check for relevant terms based on
\code{\link[=find_confounders]{find_confounders()}}. This includes the \strong{delta} argument to
set the threshold for changes in effect size that are considered relevant.
\item \strong{significant} = Check if relevant terms meet a certain threshold to be
retained in an additional model, based on \code{\link[=find_significant]{find_significant()}}.
This includes the \strong{threshold} argument that is used as a cut off value
for the \strong{statistic} that is chosen.
}}

\item{...}{Additional, optional parameters based on approach being used}
}
\value{
A \code{study} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Modify a \code{hypothesis} within a study with different approaches.
}
\concept{confounders studies}
