% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R, R/terms.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{fit}
\alias{tidy}
\alias{glance}
\alias{intersect}
\alias{union}
\alias{setdiff}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{fit}}, \code{\link[generics]{glance}}, \code{\link[generics:setops]{intersect}}, \code{\link[generics:setops]{setdiff}}, \code{\link[generics]{tidy}}, \code{\link[generics:setops]{union}}}
}}

