% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framework-helpers.R
\name{retrieval}
\alias{retrieval}
\alias{get_data}
\alias{get_test}
\alias{get_combination}
\alias{get_formulae}
\alias{get_parameters}
\alias{get_names}
\alias{update_status}
\title{Retrieve, Modify, and Destroy Components From \code{framework} Object}
\usage{
get_data(x, name)

get_test(x, name)

get_combination(x, name)

get_formulae(x, name)

get_parameters(x, name)

get_names(x, ...)

update_status(x, name, stage = NA, run = FALSE)
}
\arguments{
\item{x}{\code{framework} object}

\item{name}{Name of \code{hypothesis} object to pull components from}

\item{...}{For additional parameters to be passed on}
}
\value{
The \verb{get_*()} functions return the named object from the \code{framework}.
The \verb{update_*()} functions return the entire framework after modification.
}
\description{
These functions are essentially helper functions to simplify the manipulation
of \code{framework} objects. They may retrieve or modify or destroy components of
the \code{framework} object as described.

\strong{Retrieval functions}:
\itemize{
\item \code{get_data()} retrieves the data set associated with a specified hypothesis
and data name
\item \code{get_test()} retrieves the test associated with a hypothesis
\item \code{get_formulae()} retrieves the list of formulas associated with a
hypothesis
\item \code{get_combination()} retrieves how a hypothesis was combined to generate its
formula list
\item \code{get_names()} retrieves the hypothesis names that match a certain argument, such as \code{combination = "parallel"} or \code{data_name = "mtcars"}
}

\strong{Update functions}:
\itemize{
\item \code{update_status()} modifies the components of a framework, usually used
internally to help update the stage/status of the specified object
}
}
