% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{tbl_forest}
\alias{tbl_forest}
\title{Subgroup Forest Plot}
\usage{
tbl_forest(object, y, x, groups, levels, xlim, xbreak, xlab)
}
\arguments{
\item{object}{Class \code{forge} object with models that have been fit}
}
\description{
Subgroup Forest Plot
}
