% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt-survival.R
\name{tbl_hazard}
\alias{tbl_hazard}
\alias{tbl_dichotomous_hazard}
\alias{tbl_categorical_hazard}
\title{Make hazard ratio summary table}
\usage{
tbl_dichotomous_hazard(
  object,
  data,
  outcomes = formula(),
  followup = character(),
  terms = formula(),
  adjustment = formula(),
  rate_difference = FALSE,
  ...
)

tbl_categorical_hazard(
  object,
  data,
  outcomes = formula(),
  followup = character(),
  terms = formula(),
  adjustment = formula(),
  rate_difference = FALSE,
  ...
)
}
\arguments{
\item{object}{A \code{mdl_tbl} object with survival-based regression outcomes}

\item{data}{A dataset that corresponds to the model group of interest, and
includes survival times.}

\item{outcomes}{Is a formula or list of formulas selecting the outcome
variables of interest. The LHS is always the name of the variable that will
be selected. The RHS is the potential label for the output table. If no
label is desired, can place an \code{NA} value on RHS of formula.}

\item{followup}{Character vector naming the followup duration variable. Must
be either same length as \strong{outcomes} or be of length of 1 (which will be
recycled).}

\item{terms}{Is a formula or list of formulas selecting the model terms that
should be used. The LHS is always the name of the variable that will be
selected. The RHS is the potential label for the output table. If no label
is desired, can place an \code{NA} value on RHS of formula.}

\item{rate_difference}{If there are only two levels in the term, the rate
difference between the levels will be calculated. Defaults to \code{FALSE}.
Presumes a 95\% confidence interval as the default.}
}
\description{
Make hazard ratio summary table
}
