% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{tbl_models}
\alias{tbl_models}
\title{Make a \code{gt} table for many models}
\usage{
tbl_models(object, ...)
}
\arguments{
\item{object}{An object that represents multiple models. These should be of the same type (e.g. linear, survival, logistic, etc) for interpretability. The accepted format for this argument is as below:
\itemize{
\item \code{data.frame} An unnested/simple table that contains identifiers for multiple models that are thematically related (based on exposures, model types, etc)
\item \code{model_map} An object created from \code{\link[murmur:create_models]{murmur::create_models()}} that has been fitted using \code{\link[murmur:construct_tests]{murmur::construct_tests()}}
\item \code{list} A list of models that are related
}}

\item{...}{For passing additional arguments}

\item{x}{A variable that indicates the primary exposure. This may or may not be shared amongst models. This can be either a \emph{character vector} or a \emph{formula ~ list} object. See the \strong{details} section below for further information.}

\item{direction}{A character vector \code{c("horizontal", "vertical")}. If the
table should be oriented horizontally, with outcomes being columns, or if
the table should be oriented vertically, with outcomes being rows.}

\item{style}{If a \strong{statistic} was chosen, then specific options can be
passed to help emphasize those particular cells. This is a \emph{formula ~ list}
object where the left-hand side indicates the type of style to choose, and
the right-hand side indicates the implementation. These are based on the
[\code{\link[gt:tab_style]{gt::tab_style()}}] function.

Left-hand side should be one of \emph{fill}, \emph{text}, or \emph{borders}, which
correspond to to \code{cell_fill()}, \code{cell_text()}, and \code{cell_borders()}. The
right-hand side is a named list such as \code{list(color = lightgreen)}, and
accepts options described by the \code{gt} package. There is no validation for
the options at this time, so see documentation the \code{gt} package for
details.}
}
\description{
Make a \code{gt} table for many models
}
\details{
The \code{id}, \code{x}, \code{y}, and \code{cov} arguments can take several argument types. They
are intended to be used to help select the variable (or column), subset that
data, and re-label the data for display (if needed).
\itemize{
\item A \emph{character vector} that names either the column or location containing
the variable, or the name of the variable itself (depending on the context
of method dispatch).
\item A \emph{formula ~ list} where the LHS indicates the column or location of the
data, and the RHS indicates the variables themselves. If the RHS list is a
named list, it will use the name as the variable and the value for display.
For example, \code{term ~ list(x = "Exposure")} would represent a location
called \emph{term}, with the major variable being \emph{x}, and the displayed value
being \emph{Exposure}. This also allows subsets, as it will only retain the
variables that are listed.
}
}
\seealso{
Other visualizers: 
\code{\link{tbl_compare}()},
\code{\link{tbl_sequential}()},
\code{\link{theme_gt_compact}()}
}
\concept{visualizers}
