% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{tbls}
\alias{tbls}
\title{Extension of \code{gt} package for multiple models}
\arguments{
\item{object}{A \verb{<mdl_tbl>} object with the required models}

\item{data}{The original dataset used to create the models of interest. The
variables in the model must be contained within this dataset.}

\item{outcomes}{A \verb{<formula>} or list of formulas selecting the outcome
variables of interest. The \strong{LHS} is always the name of the variable that will
be selected. The \strong{RHS} is the potential label for the output table. If no
label is desired, can place an \code{NA} value on right-side of formula.}

\item{exposures}{A \verb{<formula>} or list of formulas of strata that should be
evaluated, with the \strong{LHS} referring to the strata and the \strong{RHS}
referring to its label.}

\item{interactions}{A \verb{<formula>} or list of formulas of the interaction
terms that should be evaluated, with the \strong{LHS} referring to the term and
the \strong{RHS} referring to its label. Currently only supports binary
interaction terms.}

\item{strata}{A \verb{<formula>} or list of formulas of strata that should be
evaluated, with the \strong{LHS} referring to the strata and the \strong{RHS}
referring to its label.}

\item{level_labels}{A \verb{<formula>} or list of formulas where each list-element
is a formula with the \strong{LHS} reflecting either the variable to re-label or
a specific level, and the \strong{RHS} reflecting what the new level should be
called (for display). If there are conflicting labels, the most recent will
be used.

For example, \verb{list(am ~ c("Manual", "Automatic")} would take, from the
\code{mtcars} dataset, the \code{am} variable, which consists of \code{c(0, 1)}, and
relabel them in the order described. They are sorted in ascending order
prior to re-labeling.

Note: If an argument has complex characters, they must be wrapped in an
additional single quotation mark to avoid issues with parsing. Hyphens (-)
and percent signs (\%) are the most common culprits.

For example, \code{speed ~ c("'>= 10\%', '< 10\%'")} would be needed to avoid an
error in parsing.}

\item{terms}{A \verb{<formula>} or list of formulas selecting the model terms that
should be used. The \strong{LHS} is always the name of the variable that will be
selected. The \strong{RHS} is the potential label for the output table. If no label
is desired, can place an \code{NA} value on right-side of formula.}
}
\description{
Extension of \code{gt} package for multiple models
}
