% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{tbls}
\alias{tbls}
\title{Extension of \code{gt} package for multiple models}
\arguments{
\item{object}{A \code{mdl_tbl} object with the required models}

\item{data}{The original dataset used to create the models of interest. The
variables in the model must be contained within this dataset.}

\item{terms}{Is a formula or list of formulas selecting the model terms that
should be used. The LHS is always the name of the variable that will be
selected. The RHS is the potential label for the output table. If no label
is desired, can place an \code{NA} value on RHS of formula.}
}
\description{
Extension of \code{gt} package for multiple models
}
